@@ -1,5 +1,7 @@
 module github.com/ory/hydra-client-go/v2
 
-go 1.13
+go 1.18
 
-require golang.org/x/oauth2 v0.0.0-20210323180902-22b0adad7558
+require (
+	golang.org/x/oauth2 v0.0.0-20210323180902-22b0adad7558
+)
