@@ -12,8 +12,13 @@ Contact: hi@ory.sh
 package openapi
 
 import (
+	"bytes"
 	"encoding/json"
+	"fmt"
 )
+
+// checks if the AcceptOAuth2LoginRequest type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &AcceptOAuth2LoginRequest{}
 
 // AcceptOAuth2LoginRequest struct for AcceptOAuth2LoginRequest
 type AcceptOAuth2LoginRequest struct {
@@ -35,6 +40,8 @@ type AcceptOAuth2LoginRequest struct {
 	Subject string `json:"subject"`
 }
 
+type _AcceptOAuth2LoginRequest AcceptOAuth2LoginRequest
+
 // NewAcceptOAuth2LoginRequest instantiates a new AcceptOAuth2LoginRequest object
 // This constructor will assign default values to properties that have it defined,
 // and makes sure properties required by API are set, but the set of arguments
@@ -55,7 +62,7 @@ func NewAcceptOAuth2LoginRequestWithDefaults() *AcceptOAuth2LoginRequest {
 
 // GetAcr returns the Acr field value if set, zero value otherwise.
 func (o *AcceptOAuth2LoginRequest) GetAcr() string {
-	if o == nil || o.Acr == nil {
+	if o == nil || IsNil(o.Acr) {
 		var ret string
 		return ret
 	}
@@ -65,7 +72,7 @@ func (o *AcceptOAuth2LoginRequest) GetAcr() string {
 // GetAcrOk returns a tuple with the Acr field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AcceptOAuth2LoginRequest) GetAcrOk() (*string, bool) {
-	if o == nil || o.Acr == nil {
+	if o == nil || IsNil(o.Acr) {
 		return nil, false
 	}
 	return o.Acr, true
@@ -73,7 +80,7 @@ func (o *AcceptOAuth2LoginRequest) GetAcrOk() (*string, bool) {
 
 // HasAcr returns a boolean if a field has been set.
 func (o *AcceptOAuth2LoginRequest) HasAcr() bool {
-	if o != nil && o.Acr != nil {
+	if o != nil && !IsNil(o.Acr) {
 		return true
 	}
 
@@ -87,7 +94,7 @@ func (o *AcceptOAuth2LoginRequest) SetAcr(v string) {
 
 // GetAmr returns the Amr field value if set, zero value otherwise.
 func (o *AcceptOAuth2LoginRequest) GetAmr() []string {
-	if o == nil || o.Amr == nil {
+	if o == nil || IsNil(o.Amr) {
 		var ret []string
 		return ret
 	}
@@ -97,7 +104,7 @@ func (o *AcceptOAuth2LoginRequest) GetAmr() []string {
 // GetAmrOk returns a tuple with the Amr field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AcceptOAuth2LoginRequest) GetAmrOk() ([]string, bool) {
-	if o == nil || o.Amr == nil {
+	if o == nil || IsNil(o.Amr) {
 		return nil, false
 	}
 	return o.Amr, true
@@ -105,7 +112,7 @@ func (o *AcceptOAuth2LoginRequest) GetAmrOk() ([]string, bool) {
 
 // HasAmr returns a boolean if a field has been set.
 func (o *AcceptOAuth2LoginRequest) HasAmr() bool {
-	if o != nil && o.Amr != nil {
+	if o != nil && !IsNil(o.Amr) {
 		return true
 	}
 
@@ -130,7 +137,7 @@ func (o *AcceptOAuth2LoginRequest) GetContext() interface{} {
 // and a boolean to check if the value has been set.
 // NOTE: If the value is an explicit nil, `nil, true` will be returned
 func (o *AcceptOAuth2LoginRequest) GetContextOk() (*interface{}, bool) {
-	if o == nil || o.Context == nil {
+	if o == nil || IsNil(o.Context) {
 		return nil, false
 	}
 	return &o.Context, true
@@ -138,7 +145,7 @@ func (o *AcceptOAuth2LoginRequest) GetContextOk() (*interface{}, bool) {
 
 // HasContext returns a boolean if a field has been set.
 func (o *AcceptOAuth2LoginRequest) HasContext() bool {
-	if o != nil && o.Context != nil {
+	if o != nil && IsNil(o.Context) {
 		return true
 	}
 
@@ -152,7 +159,7 @@ func (o *AcceptOAuth2LoginRequest) SetContext(v interface{}) {
 
 // GetExtendSessionLifespan returns the ExtendSessionLifespan field value if set, zero value otherwise.
 func (o *AcceptOAuth2LoginRequest) GetExtendSessionLifespan() bool {
-	if o == nil || o.ExtendSessionLifespan == nil {
+	if o == nil || IsNil(o.ExtendSessionLifespan) {
 		var ret bool
 		return ret
 	}
@@ -162,7 +169,7 @@ func (o *AcceptOAuth2LoginRequest) GetExtendSessionLifespan() bool {
 // GetExtendSessionLifespanOk returns a tuple with the ExtendSessionLifespan field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AcceptOAuth2LoginRequest) GetExtendSessionLifespanOk() (*bool, bool) {
-	if o == nil || o.ExtendSessionLifespan == nil {
+	if o == nil || IsNil(o.ExtendSessionLifespan) {
 		return nil, false
 	}
 	return o.ExtendSessionLifespan, true
@@ -170,7 +177,7 @@ func (o *AcceptOAuth2LoginRequest) GetExtendSessionLifespanOk() (*bool, bool) {
 
 // HasExtendSessionLifespan returns a boolean if a field has been set.
 func (o *AcceptOAuth2LoginRequest) HasExtendSessionLifespan() bool {
-	if o != nil && o.ExtendSessionLifespan != nil {
+	if o != nil && !IsNil(o.ExtendSessionLifespan) {
 		return true
 	}
 
@@ -184,7 +191,7 @@ func (o *AcceptOAuth2LoginRequest) SetExtendSessionLifespan(v bool) {
 
 // GetForceSubjectIdentifier returns the ForceSubjectIdentifier field value if set, zero value otherwise.
 func (o *AcceptOAuth2LoginRequest) GetForceSubjectIdentifier() string {
-	if o == nil || o.ForceSubjectIdentifier == nil {
+	if o == nil || IsNil(o.ForceSubjectIdentifier) {
 		var ret string
 		return ret
 	}
@@ -194,7 +201,7 @@ func (o *AcceptOAuth2LoginRequest) GetForceSubjectIdentifier() string {
 // GetForceSubjectIdentifierOk returns a tuple with the ForceSubjectIdentifier field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AcceptOAuth2LoginRequest) GetForceSubjectIdentifierOk() (*string, bool) {
-	if o == nil || o.ForceSubjectIdentifier == nil {
+	if o == nil || IsNil(o.ForceSubjectIdentifier) {
 		return nil, false
 	}
 	return o.ForceSubjectIdentifier, true
@@ -202,7 +209,7 @@ func (o *AcceptOAuth2LoginRequest) GetForceSubjectIdentifierOk() (*string, bool)
 
 // HasForceSubjectIdentifier returns a boolean if a field has been set.
 func (o *AcceptOAuth2LoginRequest) HasForceSubjectIdentifier() bool {
-	if o != nil && o.ForceSubjectIdentifier != nil {
+	if o != nil && !IsNil(o.ForceSubjectIdentifier) {
 		return true
 	}
 
@@ -216,7 +223,7 @@ func (o *AcceptOAuth2LoginRequest) SetForceSubjectIdentifier(v string) {
 
 // GetIdentityProviderSessionId returns the IdentityProviderSessionId field value if set, zero value otherwise.
 func (o *AcceptOAuth2LoginRequest) GetIdentityProviderSessionId() string {
-	if o == nil || o.IdentityProviderSessionId == nil {
+	if o == nil || IsNil(o.IdentityProviderSessionId) {
 		var ret string
 		return ret
 	}
@@ -226,7 +233,7 @@ func (o *AcceptOAuth2LoginRequest) GetIdentityProviderSessionId() string {
 // GetIdentityProviderSessionIdOk returns a tuple with the IdentityProviderSessionId field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AcceptOAuth2LoginRequest) GetIdentityProviderSessionIdOk() (*string, bool) {
-	if o == nil || o.IdentityProviderSessionId == nil {
+	if o == nil || IsNil(o.IdentityProviderSessionId) {
 		return nil, false
 	}
 	return o.IdentityProviderSessionId, true
@@ -234,7 +241,7 @@ func (o *AcceptOAuth2LoginRequest) GetIdentityProviderSessionIdOk() (*string, bo
 
 // HasIdentityProviderSessionId returns a boolean if a field has been set.
 func (o *AcceptOAuth2LoginRequest) HasIdentityProviderSessionId() bool {
-	if o != nil && o.IdentityProviderSessionId != nil {
+	if o != nil && !IsNil(o.IdentityProviderSessionId) {
 		return true
 	}
 
@@ -248,7 +255,7 @@ func (o *AcceptOAuth2LoginRequest) SetIdentityProviderSessionId(v string) {
 
 // GetRemember returns the Remember field value if set, zero value otherwise.
 func (o *AcceptOAuth2LoginRequest) GetRemember() bool {
-	if o == nil || o.Remember == nil {
+	if o == nil || IsNil(o.Remember) {
 		var ret bool
 		return ret
 	}
@@ -258,7 +265,7 @@ func (o *AcceptOAuth2LoginRequest) GetRemember() bool {
 // GetRememberOk returns a tuple with the Remember field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AcceptOAuth2LoginRequest) GetRememberOk() (*bool, bool) {
-	if o == nil || o.Remember == nil {
+	if o == nil || IsNil(o.Remember) {
 		return nil, false
 	}
 	return o.Remember, true
@@ -266,7 +273,7 @@ func (o *AcceptOAuth2LoginRequest) GetRememberOk() (*bool, bool) {
 
 // HasRemember returns a boolean if a field has been set.
 func (o *AcceptOAuth2LoginRequest) HasRemember() bool {
-	if o != nil && o.Remember != nil {
+	if o != nil && !IsNil(o.Remember) {
 		return true
 	}
 
@@ -280,7 +287,7 @@ func (o *AcceptOAuth2LoginRequest) SetRemember(v bool) {
 
 // GetRememberFor returns the RememberFor field value if set, zero value otherwise.
 func (o *AcceptOAuth2LoginRequest) GetRememberFor() int64 {
-	if o == nil || o.RememberFor == nil {
+	if o == nil || IsNil(o.RememberFor) {
 		var ret int64
 		return ret
 	}
@@ -290,7 +297,7 @@ func (o *AcceptOAuth2LoginRequest) GetRememberFor() int64 {
 // GetRememberForOk returns a tuple with the RememberFor field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AcceptOAuth2LoginRequest) GetRememberForOk() (*int64, bool) {
-	if o == nil || o.RememberFor == nil {
+	if o == nil || IsNil(o.RememberFor) {
 		return nil, false
 	}
 	return o.RememberFor, true
@@ -298,7 +305,7 @@ func (o *AcceptOAuth2LoginRequest) GetRememberForOk() (*int64, bool) {
 
 // HasRememberFor returns a boolean if a field has been set.
 func (o *AcceptOAuth2LoginRequest) HasRememberFor() bool {
-	if o != nil && o.RememberFor != nil {
+	if o != nil && !IsNil(o.RememberFor) {
 		return true
 	}
 
@@ -335,35 +342,78 @@ func (o *AcceptOAuth2LoginRequest) SetSubject(v string) {
 }
 
 func (o AcceptOAuth2LoginRequest) MarshalJSON() ([]byte, error) {
+	toSerialize, err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
+	}
+	return json.Marshal(toSerialize)
+}
+
+func (o AcceptOAuth2LoginRequest) ToMap() (map[string]interface{}, error) {
 	toSerialize := map[string]interface{}{}
-	if o.Acr != nil {
+	if !IsNil(o.Acr) {
 		toSerialize["acr"] = o.Acr
 	}
-	if o.Amr != nil {
+	if !IsNil(o.Amr) {
 		toSerialize["amr"] = o.Amr
 	}
 	if o.Context != nil {
 		toSerialize["context"] = o.Context
 	}
-	if o.ExtendSessionLifespan != nil {
+	if !IsNil(o.ExtendSessionLifespan) {
 		toSerialize["extend_session_lifespan"] = o.ExtendSessionLifespan
 	}
-	if o.ForceSubjectIdentifier != nil {
+	if !IsNil(o.ForceSubjectIdentifier) {
 		toSerialize["force_subject_identifier"] = o.ForceSubjectIdentifier
 	}
-	if o.IdentityProviderSessionId != nil {
+	if !IsNil(o.IdentityProviderSessionId) {
 		toSerialize["identity_provider_session_id"] = o.IdentityProviderSessionId
 	}
-	if o.Remember != nil {
+	if !IsNil(o.Remember) {
 		toSerialize["remember"] = o.Remember
 	}
-	if o.RememberFor != nil {
+	if !IsNil(o.RememberFor) {
 		toSerialize["remember_for"] = o.RememberFor
 	}
-	if true {
-		toSerialize["subject"] = o.Subject
+	toSerialize["subject"] = o.Subject
+	return toSerialize, nil
+}
+
+func (o *AcceptOAuth2LoginRequest) UnmarshalJSON(data []byte) (err error) {
+	// This validates that all required properties are included in the JSON object
+	// by unmarshalling the object into a generic map with string keys and checking
+	// that every required field exists as a key in the generic map.
+	requiredProperties := []string{
+		"subject",
 	}
-	return json.Marshal(toSerialize)
+
+	allProperties := make(map[string]interface{})
+
+	err = json.Unmarshal(data, &allProperties)
+
+	if err != nil {
+		return err
+	}
+
+	for _, requiredProperty := range requiredProperties {
+		if _, exists := allProperties[requiredProperty]; !exists {
+			return fmt.Errorf("no value given for required property %v", requiredProperty)
+		}
+	}
+
+	varAcceptOAuth2LoginRequest := _AcceptOAuth2LoginRequest{}
+
+	decoder := json.NewDecoder(bytes.NewReader(data))
+	decoder.DisallowUnknownFields()
+	err = decoder.Decode(&varAcceptOAuth2LoginRequest)
+
+	if err != nil {
+		return err
+	}
+
+	*o = AcceptOAuth2LoginRequest(varAcceptOAuth2LoginRequest)
+
+	return err
 }
 
 type NullableAcceptOAuth2LoginRequest struct {
