@@ -12,8 +12,13 @@ Contact: hi@ory.sh
 package openapi
 
 import (
+	"bytes"
 	"encoding/json"
+	"fmt"
 )
+
+// checks if the CreateJsonWebKeySet type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &CreateJsonWebKeySet{}
 
 // CreateJsonWebKeySet Create JSON Web Key Set Request Body
 type CreateJsonWebKeySet struct {
@@ -24,6 +29,8 @@ type CreateJsonWebKeySet struct {
 	// JSON Web Key Use  The \"use\" (public key use) parameter identifies the intended use of the public key. The \"use\" parameter is employed to indicate whether a public key is used for encrypting data or verifying the signature on data. Valid values are \"enc\" and \"sig\".
 	Use string `json:"use"`
 }
+
+type _CreateJsonWebKeySet CreateJsonWebKeySet
 
 // NewCreateJsonWebKeySet instantiates a new CreateJsonWebKeySet object
 // This constructor will assign default values to properties that have it defined,
@@ -118,17 +125,58 @@ func (o *CreateJsonWebKeySet) SetUse(v string) {
 }
 
 func (o CreateJsonWebKeySet) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if true {
-		toSerialize["alg"] = o.Alg
-	}
-	if true {
-		toSerialize["kid"] = o.Kid
-	}
-	if true {
-		toSerialize["use"] = o.Use
+	toSerialize, err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o CreateJsonWebKeySet) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	toSerialize["alg"] = o.Alg
+	toSerialize["kid"] = o.Kid
+	toSerialize["use"] = o.Use
+	return toSerialize, nil
+}
+
+func (o *CreateJsonWebKeySet) UnmarshalJSON(data []byte) (err error) {
+	// This validates that all required properties are included in the JSON object
+	// by unmarshalling the object into a generic map with string keys and checking
+	// that every required field exists as a key in the generic map.
+	requiredProperties := []string{
+		"alg",
+		"kid",
+		"use",
+	}
+
+	allProperties := make(map[string]interface{})
+
+	err = json.Unmarshal(data, &allProperties)
+
+	if err != nil {
+		return err
+	}
+
+	for _, requiredProperty := range requiredProperties {
+		if _, exists := allProperties[requiredProperty]; !exists {
+			return fmt.Errorf("no value given for required property %v", requiredProperty)
+		}
+	}
+
+	varCreateJsonWebKeySet := _CreateJsonWebKeySet{}
+
+	decoder := json.NewDecoder(bytes.NewReader(data))
+	decoder.DisallowUnknownFields()
+	err = decoder.Decode(&varCreateJsonWebKeySet)
+
+	if err != nil {
+		return err
+	}
+
+	*o = CreateJsonWebKeySet(varCreateJsonWebKeySet)
+
+	return err
 }
 
 type NullableCreateJsonWebKeySet struct {
