@@ -12,8 +12,13 @@ Contact: hi@ory.sh
 package openapi
 
 import (
+	"bytes"
 	"encoding/json"
+	"fmt"
 )
+
+// checks if the GenericError type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &GenericError{}
 
 // GenericError struct for GenericError
 type GenericError struct {
@@ -35,6 +40,8 @@ type GenericError struct {
 	Status *string `json:"status,omitempty"`
 }
 
+type _GenericError GenericError
+
 // NewGenericError instantiates a new GenericError object
 // This constructor will assign default values to properties that have it defined,
 // and makes sure properties required by API are set, but the set of arguments
@@ -55,7 +62,7 @@ func NewGenericErrorWithDefaults() *GenericError {
 
 // GetCode returns the Code field value if set, zero value otherwise.
 func (o *GenericError) GetCode() int64 {
-	if o == nil || o.Code == nil {
+	if o == nil || IsNil(o.Code) {
 		var ret int64
 		return ret
 	}
@@ -65,7 +72,7 @@ func (o *GenericError) GetCode() int64 {
 // GetCodeOk returns a tuple with the Code field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *GenericError) GetCodeOk() (*int64, bool) {
-	if o == nil || o.Code == nil {
+	if o == nil || IsNil(o.Code) {
 		return nil, false
 	}
 	return o.Code, true
@@ -73,7 +80,7 @@ func (o *GenericError) GetCodeOk() (*int64, bool) {
 
 // HasCode returns a boolean if a field has been set.
 func (o *GenericError) HasCode() bool {
-	if o != nil && o.Code != nil {
+	if o != nil && !IsNil(o.Code) {
 		return true
 	}
 
@@ -87,7 +94,7 @@ func (o *GenericError) SetCode(v int64) {
 
 // GetDebug returns the Debug field value if set, zero value otherwise.
 func (o *GenericError) GetDebug() string {
-	if o == nil || o.Debug == nil {
+	if o == nil || IsNil(o.Debug) {
 		var ret string
 		return ret
 	}
@@ -97,7 +104,7 @@ func (o *GenericError) GetDebug() string {
 // GetDebugOk returns a tuple with the Debug field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *GenericError) GetDebugOk() (*string, bool) {
-	if o == nil || o.Debug == nil {
+	if o == nil || IsNil(o.Debug) {
 		return nil, false
 	}
 	return o.Debug, true
@@ -105,7 +112,7 @@ func (o *GenericError) GetDebugOk() (*string, bool) {
 
 // HasDebug returns a boolean if a field has been set.
 func (o *GenericError) HasDebug() bool {
-	if o != nil && o.Debug != nil {
+	if o != nil && !IsNil(o.Debug) {
 		return true
 	}
 
@@ -130,7 +137,7 @@ func (o *GenericError) GetDetails() interface{} {
 // and a boolean to check if the value has been set.
 // NOTE: If the value is an explicit nil, `nil, true` will be returned
 func (o *GenericError) GetDetailsOk() (*interface{}, bool) {
-	if o == nil || o.Details == nil {
+	if o == nil || IsNil(o.Details) {
 		return nil, false
 	}
 	return &o.Details, true
@@ -138,7 +145,7 @@ func (o *GenericError) GetDetailsOk() (*interface{}, bool) {
 
 // HasDetails returns a boolean if a field has been set.
 func (o *GenericError) HasDetails() bool {
-	if o != nil && o.Details != nil {
+	if o != nil && IsNil(o.Details) {
 		return true
 	}
 
@@ -152,7 +159,7 @@ func (o *GenericError) SetDetails(v interface{}) {
 
 // GetId returns the Id field value if set, zero value otherwise.
 func (o *GenericError) GetId() string {
-	if o == nil || o.Id == nil {
+	if o == nil || IsNil(o.Id) {
 		var ret string
 		return ret
 	}
@@ -162,7 +169,7 @@ func (o *GenericError) GetId() string {
 // GetIdOk returns a tuple with the Id field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *GenericError) GetIdOk() (*string, bool) {
-	if o == nil || o.Id == nil {
+	if o == nil || IsNil(o.Id) {
 		return nil, false
 	}
 	return o.Id, true
@@ -170,7 +177,7 @@ func (o *GenericError) GetIdOk() (*string, bool) {
 
 // HasId returns a boolean if a field has been set.
 func (o *GenericError) HasId() bool {
-	if o != nil && o.Id != nil {
+	if o != nil && !IsNil(o.Id) {
 		return true
 	}
 
@@ -208,7 +215,7 @@ func (o *GenericError) SetMessage(v string) {
 
 // GetReason returns the Reason field value if set, zero value otherwise.
 func (o *GenericError) GetReason() string {
-	if o == nil || o.Reason == nil {
+	if o == nil || IsNil(o.Reason) {
 		var ret string
 		return ret
 	}
@@ -218,7 +225,7 @@ func (o *GenericError) GetReason() string {
 // GetReasonOk returns a tuple with the Reason field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *GenericError) GetReasonOk() (*string, bool) {
-	if o == nil || o.Reason == nil {
+	if o == nil || IsNil(o.Reason) {
 		return nil, false
 	}
 	return o.Reason, true
@@ -226,7 +233,7 @@ func (o *GenericError) GetReasonOk() (*string, bool) {
 
 // HasReason returns a boolean if a field has been set.
 func (o *GenericError) HasReason() bool {
-	if o != nil && o.Reason != nil {
+	if o != nil && !IsNil(o.Reason) {
 		return true
 	}
 
@@ -240,7 +247,7 @@ func (o *GenericError) SetReason(v string) {
 
 // GetRequest returns the Request field value if set, zero value otherwise.
 func (o *GenericError) GetRequest() string {
-	if o == nil || o.Request == nil {
+	if o == nil || IsNil(o.Request) {
 		var ret string
 		return ret
 	}
@@ -250,7 +257,7 @@ func (o *GenericError) GetRequest() string {
 // GetRequestOk returns a tuple with the Request field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *GenericError) GetRequestOk() (*string, bool) {
-	if o == nil || o.Request == nil {
+	if o == nil || IsNil(o.Request) {
 		return nil, false
 	}
 	return o.Request, true
@@ -258,7 +265,7 @@ func (o *GenericError) GetRequestOk() (*string, bool) {
 
 // HasRequest returns a boolean if a field has been set.
 func (o *GenericError) HasRequest() bool {
-	if o != nil && o.Request != nil {
+	if o != nil && !IsNil(o.Request) {
 		return true
 	}
 
@@ -272,7 +279,7 @@ func (o *GenericError) SetRequest(v string) {
 
 // GetStatus returns the Status field value if set, zero value otherwise.
 func (o *GenericError) GetStatus() string {
-	if o == nil || o.Status == nil {
+	if o == nil || IsNil(o.Status) {
 		var ret string
 		return ret
 	}
@@ -282,7 +289,7 @@ func (o *GenericError) GetStatus() string {
 // GetStatusOk returns a tuple with the Status field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *GenericError) GetStatusOk() (*string, bool) {
-	if o == nil || o.Status == nil {
+	if o == nil || IsNil(o.Status) {
 		return nil, false
 	}
 	return o.Status, true
@@ -290,7 +297,7 @@ func (o *GenericError) GetStatusOk() (*string, bool) {
 
 // HasStatus returns a boolean if a field has been set.
 func (o *GenericError) HasStatus() bool {
-	if o != nil && o.Status != nil {
+	if o != nil && !IsNil(o.Status) {
 		return true
 	}
 
@@ -303,32 +310,75 @@ func (o *GenericError) SetStatus(v string) {
 }
 
 func (o GenericError) MarshalJSON() ([]byte, error) {
+	toSerialize, err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
+	}
+	return json.Marshal(toSerialize)
+}
+
+func (o GenericError) ToMap() (map[string]interface{}, error) {
 	toSerialize := map[string]interface{}{}
-	if o.Code != nil {
+	if !IsNil(o.Code) {
 		toSerialize["code"] = o.Code
 	}
-	if o.Debug != nil {
+	if !IsNil(o.Debug) {
 		toSerialize["debug"] = o.Debug
 	}
 	if o.Details != nil {
 		toSerialize["details"] = o.Details
 	}
-	if o.Id != nil {
+	if !IsNil(o.Id) {
 		toSerialize["id"] = o.Id
 	}
-	if true {
-		toSerialize["message"] = o.Message
-	}
-	if o.Reason != nil {
+	toSerialize["message"] = o.Message
+	if !IsNil(o.Reason) {
 		toSerialize["reason"] = o.Reason
 	}
-	if o.Request != nil {
+	if !IsNil(o.Request) {
 		toSerialize["request"] = o.Request
 	}
-	if o.Status != nil {
+	if !IsNil(o.Status) {
 		toSerialize["status"] = o.Status
 	}
-	return json.Marshal(toSerialize)
+	return toSerialize, nil
+}
+
+func (o *GenericError) UnmarshalJSON(data []byte) (err error) {
+	// This validates that all required properties are included in the JSON object
+	// by unmarshalling the object into a generic map with string keys and checking
+	// that every required field exists as a key in the generic map.
+	requiredProperties := []string{
+		"message",
+	}
+
+	allProperties := make(map[string]interface{})
+
+	err = json.Unmarshal(data, &allProperties)
+
+	if err != nil {
+		return err
+	}
+
+	for _, requiredProperty := range requiredProperties {
+		if _, exists := allProperties[requiredProperty]; !exists {
+			return fmt.Errorf("no value given for required property %v", requiredProperty)
+		}
+	}
+
+	varGenericError := _GenericError{}
+
+	decoder := json.NewDecoder(bytes.NewReader(data))
+	decoder.DisallowUnknownFields()
+	err = decoder.Decode(&varGenericError)
+
+	if err != nil {
+		return err
+	}
+
+	*o = GenericError(varGenericError)
+
+	return err
 }
 
 type NullableGenericError struct {
