@@ -15,6 +15,9 @@ import (
 	"encoding/json"
 )
 
+// checks if the GetVersion200Response type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &GetVersion200Response{}
+
 // GetVersion200Response struct for GetVersion200Response
 type GetVersion200Response struct {
 	// The version of Ory Hydra.
@@ -40,7 +43,7 @@ func NewGetVersion200ResponseWithDefaults() *GetVersion200Response {
 
 // GetVersion returns the Version field value if set, zero value otherwise.
 func (o *GetVersion200Response) GetVersion() string {
-	if o == nil || o.Version == nil {
+	if o == nil || IsNil(o.Version) {
 		var ret string
 		return ret
 	}
@@ -50,7 +53,7 @@ func (o *GetVersion200Response) GetVersion() string {
 // GetVersionOk returns a tuple with the Version field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *GetVersion200Response) GetVersionOk() (*string, bool) {
-	if o == nil || o.Version == nil {
+	if o == nil || IsNil(o.Version) {
 		return nil, false
 	}
 	return o.Version, true
@@ -58,7 +61,7 @@ func (o *GetVersion200Response) GetVersionOk() (*string, bool) {
 
 // HasVersion returns a boolean if a field has been set.
 func (o *GetVersion200Response) HasVersion() bool {
-	if o != nil && o.Version != nil {
+	if o != nil && !IsNil(o.Version) {
 		return true
 	}
 
@@ -71,11 +74,19 @@ func (o *GetVersion200Response) SetVersion(v string) {
 }
 
 func (o GetVersion200Response) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.Version != nil {
-		toSerialize["version"] = o.Version
+	toSerialize, err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o GetVersion200Response) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.Version) {
+		toSerialize["version"] = o.Version
+	}
+	return toSerialize, nil
 }
 
 type NullableGetVersion200Response struct {
