@@ -15,6 +15,9 @@ import (
 	"encoding/json"
 )
 
+// checks if the HealthNotReadyStatus type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &HealthNotReadyStatus{}
+
 // HealthNotReadyStatus struct for HealthNotReadyStatus
 type HealthNotReadyStatus struct {
 	// Errors contains a list of errors that caused the not ready status.
@@ -40,7 +43,7 @@ func NewHealthNotReadyStatusWithDefaults() *HealthNotReadyStatus {
 
 // GetErrors returns the Errors field value if set, zero value otherwise.
 func (o *HealthNotReadyStatus) GetErrors() map[string]string {
-	if o == nil || o.Errors == nil {
+	if o == nil || IsNil(o.Errors) {
 		var ret map[string]string
 		return ret
 	}
@@ -50,7 +53,7 @@ func (o *HealthNotReadyStatus) GetErrors() map[string]string {
 // GetErrorsOk returns a tuple with the Errors field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *HealthNotReadyStatus) GetErrorsOk() (*map[string]string, bool) {
-	if o == nil || o.Errors == nil {
+	if o == nil || IsNil(o.Errors) {
 		return nil, false
 	}
 	return o.Errors, true
@@ -58,7 +61,7 @@ func (o *HealthNotReadyStatus) GetErrorsOk() (*map[string]string, bool) {
 
 // HasErrors returns a boolean if a field has been set.
 func (o *HealthNotReadyStatus) HasErrors() bool {
-	if o != nil && o.Errors != nil {
+	if o != nil && !IsNil(o.Errors) {
 		return true
 	}
 
@@ -71,11 +74,19 @@ func (o *HealthNotReadyStatus) SetErrors(v map[string]string) {
 }
 
 func (o HealthNotReadyStatus) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.Errors != nil {
-		toSerialize["errors"] = o.Errors
+	toSerialize, err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o HealthNotReadyStatus) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.Errors) {
+		toSerialize["errors"] = o.Errors
+	}
+	return toSerialize, nil
 }
 
 type NullableHealthNotReadyStatus struct {
