@@ -12,8 +12,13 @@ Contact: hi@ory.sh
 package openapi
 
 import (
+	"bytes"
 	"encoding/json"
+	"fmt"
 )
+
+// checks if the IntrospectedOAuth2Token type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &IntrospectedOAuth2Token{}
 
 // IntrospectedOAuth2Token Introspection contains an access token's session data as specified by [IETF RFC 7662](https://tools.ietf.org/html/rfc7662)
 type IntrospectedOAuth2Token struct {
@@ -46,6 +51,8 @@ type IntrospectedOAuth2Token struct {
 	// Username is a human-readable identifier for the resource owner who authorized this token.
 	Username *string `json:"username,omitempty"`
 }
+
+type _IntrospectedOAuth2Token IntrospectedOAuth2Token
 
 // NewIntrospectedOAuth2Token instantiates a new IntrospectedOAuth2Token object
 // This constructor will assign default values to properties that have it defined,
@@ -91,7 +98,7 @@ func (o *IntrospectedOAuth2Token) SetActive(v bool) {
 
 // GetAud returns the Aud field value if set, zero value otherwise.
 func (o *IntrospectedOAuth2Token) GetAud() []string {
-	if o == nil || o.Aud == nil {
+	if o == nil || IsNil(o.Aud) {
 		var ret []string
 		return ret
 	}
@@ -101,7 +108,7 @@ func (o *IntrospectedOAuth2Token) GetAud() []string {
 // GetAudOk returns a tuple with the Aud field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *IntrospectedOAuth2Token) GetAudOk() ([]string, bool) {
-	if o == nil || o.Aud == nil {
+	if o == nil || IsNil(o.Aud) {
 		return nil, false
 	}
 	return o.Aud, true
@@ -109,7 +116,7 @@ func (o *IntrospectedOAuth2Token) GetAudOk() ([]string, bool) {
 
 // HasAud returns a boolean if a field has been set.
 func (o *IntrospectedOAuth2Token) HasAud() bool {
-	if o != nil && o.Aud != nil {
+	if o != nil && !IsNil(o.Aud) {
 		return true
 	}
 
@@ -123,7 +130,7 @@ func (o *IntrospectedOAuth2Token) SetAud(v []string) {
 
 // GetClientId returns the ClientId field value if set, zero value otherwise.
 func (o *IntrospectedOAuth2Token) GetClientId() string {
-	if o == nil || o.ClientId == nil {
+	if o == nil || IsNil(o.ClientId) {
 		var ret string
 		return ret
 	}
@@ -133,7 +140,7 @@ func (o *IntrospectedOAuth2Token) GetClientId() string {
 // GetClientIdOk returns a tuple with the ClientId field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *IntrospectedOAuth2Token) GetClientIdOk() (*string, bool) {
-	if o == nil || o.ClientId == nil {
+	if o == nil || IsNil(o.ClientId) {
 		return nil, false
 	}
 	return o.ClientId, true
@@ -141,7 +148,7 @@ func (o *IntrospectedOAuth2Token) GetClientIdOk() (*string, bool) {
 
 // HasClientId returns a boolean if a field has been set.
 func (o *IntrospectedOAuth2Token) HasClientId() bool {
-	if o != nil && o.ClientId != nil {
+	if o != nil && !IsNil(o.ClientId) {
 		return true
 	}
 
@@ -155,7 +162,7 @@ func (o *IntrospectedOAuth2Token) SetClientId(v string) {
 
 // GetExp returns the Exp field value if set, zero value otherwise.
 func (o *IntrospectedOAuth2Token) GetExp() int64 {
-	if o == nil || o.Exp == nil {
+	if o == nil || IsNil(o.Exp) {
 		var ret int64
 		return ret
 	}
@@ -165,7 +172,7 @@ func (o *IntrospectedOAuth2Token) GetExp() int64 {
 // GetExpOk returns a tuple with the Exp field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *IntrospectedOAuth2Token) GetExpOk() (*int64, bool) {
-	if o == nil || o.Exp == nil {
+	if o == nil || IsNil(o.Exp) {
 		return nil, false
 	}
 	return o.Exp, true
@@ -173,7 +180,7 @@ func (o *IntrospectedOAuth2Token) GetExpOk() (*int64, bool) {
 
 // HasExp returns a boolean if a field has been set.
 func (o *IntrospectedOAuth2Token) HasExp() bool {
-	if o != nil && o.Exp != nil {
+	if o != nil && !IsNil(o.Exp) {
 		return true
 	}
 
@@ -187,7 +194,7 @@ func (o *IntrospectedOAuth2Token) SetExp(v int64) {
 
 // GetExt returns the Ext field value if set, zero value otherwise.
 func (o *IntrospectedOAuth2Token) GetExt() map[string]interface{} {
-	if o == nil || o.Ext == nil {
+	if o == nil || IsNil(o.Ext) {
 		var ret map[string]interface{}
 		return ret
 	}
@@ -197,15 +204,15 @@ func (o *IntrospectedOAuth2Token) GetExt() map[string]interface{} {
 // GetExtOk returns a tuple with the Ext field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *IntrospectedOAuth2Token) GetExtOk() (map[string]interface{}, bool) {
-	if o == nil || o.Ext == nil {
-		return nil, false
+	if o == nil || IsNil(o.Ext) {
+		return map[string]interface{}{}, false
 	}
 	return o.Ext, true
 }
 
 // HasExt returns a boolean if a field has been set.
 func (o *IntrospectedOAuth2Token) HasExt() bool {
-	if o != nil && o.Ext != nil {
+	if o != nil && !IsNil(o.Ext) {
 		return true
 	}
 
@@ -219,7 +226,7 @@ func (o *IntrospectedOAuth2Token) SetExt(v map[string]interface{}) {
 
 // GetIat returns the Iat field value if set, zero value otherwise.
 func (o *IntrospectedOAuth2Token) GetIat() int64 {
-	if o == nil || o.Iat == nil {
+	if o == nil || IsNil(o.Iat) {
 		var ret int64
 		return ret
 	}
@@ -229,7 +236,7 @@ func (o *IntrospectedOAuth2Token) GetIat() int64 {
 // GetIatOk returns a tuple with the Iat field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *IntrospectedOAuth2Token) GetIatOk() (*int64, bool) {
-	if o == nil || o.Iat == nil {
+	if o == nil || IsNil(o.Iat) {
 		return nil, false
 	}
 	return o.Iat, true
@@ -237,7 +244,7 @@ func (o *IntrospectedOAuth2Token) GetIatOk() (*int64, bool) {
 
 // HasIat returns a boolean if a field has been set.
 func (o *IntrospectedOAuth2Token) HasIat() bool {
-	if o != nil && o.Iat != nil {
+	if o != nil && !IsNil(o.Iat) {
 		return true
 	}
 
@@ -251,7 +258,7 @@ func (o *IntrospectedOAuth2Token) SetIat(v int64) {
 
 // GetIss returns the Iss field value if set, zero value otherwise.
 func (o *IntrospectedOAuth2Token) GetIss() string {
-	if o == nil || o.Iss == nil {
+	if o == nil || IsNil(o.Iss) {
 		var ret string
 		return ret
 	}
@@ -261,7 +268,7 @@ func (o *IntrospectedOAuth2Token) GetIss() string {
 // GetIssOk returns a tuple with the Iss field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *IntrospectedOAuth2Token) GetIssOk() (*string, bool) {
-	if o == nil || o.Iss == nil {
+	if o == nil || IsNil(o.Iss) {
 		return nil, false
 	}
 	return o.Iss, true
@@ -269,7 +276,7 @@ func (o *IntrospectedOAuth2Token) GetIssOk() (*string, bool) {
 
 // HasIss returns a boolean if a field has been set.
 func (o *IntrospectedOAuth2Token) HasIss() bool {
-	if o != nil && o.Iss != nil {
+	if o != nil && !IsNil(o.Iss) {
 		return true
 	}
 
@@ -283,7 +290,7 @@ func (o *IntrospectedOAuth2Token) SetIss(v string) {
 
 // GetNbf returns the Nbf field value if set, zero value otherwise.
 func (o *IntrospectedOAuth2Token) GetNbf() int64 {
-	if o == nil || o.Nbf == nil {
+	if o == nil || IsNil(o.Nbf) {
 		var ret int64
 		return ret
 	}
@@ -293,7 +300,7 @@ func (o *IntrospectedOAuth2Token) GetNbf() int64 {
 // GetNbfOk returns a tuple with the Nbf field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *IntrospectedOAuth2Token) GetNbfOk() (*int64, bool) {
-	if o == nil || o.Nbf == nil {
+	if o == nil || IsNil(o.Nbf) {
 		return nil, false
 	}
 	return o.Nbf, true
@@ -301,7 +308,7 @@ func (o *IntrospectedOAuth2Token) GetNbfOk() (*int64, bool) {
 
 // HasNbf returns a boolean if a field has been set.
 func (o *IntrospectedOAuth2Token) HasNbf() bool {
-	if o != nil && o.Nbf != nil {
+	if o != nil && !IsNil(o.Nbf) {
 		return true
 	}
 
@@ -315,7 +322,7 @@ func (o *IntrospectedOAuth2Token) SetNbf(v int64) {
 
 // GetObfuscatedSubject returns the ObfuscatedSubject field value if set, zero value otherwise.
 func (o *IntrospectedOAuth2Token) GetObfuscatedSubject() string {
-	if o == nil || o.ObfuscatedSubject == nil {
+	if o == nil || IsNil(o.ObfuscatedSubject) {
 		var ret string
 		return ret
 	}
@@ -325,7 +332,7 @@ func (o *IntrospectedOAuth2Token) GetObfuscatedSubject() string {
 // GetObfuscatedSubjectOk returns a tuple with the ObfuscatedSubject field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *IntrospectedOAuth2Token) GetObfuscatedSubjectOk() (*string, bool) {
-	if o == nil || o.ObfuscatedSubject == nil {
+	if o == nil || IsNil(o.ObfuscatedSubject) {
 		return nil, false
 	}
 	return o.ObfuscatedSubject, true
@@ -333,7 +340,7 @@ func (o *IntrospectedOAuth2Token) GetObfuscatedSubjectOk() (*string, bool) {
 
 // HasObfuscatedSubject returns a boolean if a field has been set.
 func (o *IntrospectedOAuth2Token) HasObfuscatedSubject() bool {
-	if o != nil && o.ObfuscatedSubject != nil {
+	if o != nil && !IsNil(o.ObfuscatedSubject) {
 		return true
 	}
 
@@ -347,7 +354,7 @@ func (o *IntrospectedOAuth2Token) SetObfuscatedSubject(v string) {
 
 // GetScope returns the Scope field value if set, zero value otherwise.
 func (o *IntrospectedOAuth2Token) GetScope() string {
-	if o == nil || o.Scope == nil {
+	if o == nil || IsNil(o.Scope) {
 		var ret string
 		return ret
 	}
@@ -357,7 +364,7 @@ func (o *IntrospectedOAuth2Token) GetScope() string {
 // GetScopeOk returns a tuple with the Scope field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *IntrospectedOAuth2Token) GetScopeOk() (*string, bool) {
-	if o == nil || o.Scope == nil {
+	if o == nil || IsNil(o.Scope) {
 		return nil, false
 	}
 	return o.Scope, true
@@ -365,7 +372,7 @@ func (o *IntrospectedOAuth2Token) GetScopeOk() (*string, bool) {
 
 // HasScope returns a boolean if a field has been set.
 func (o *IntrospectedOAuth2Token) HasScope() bool {
-	if o != nil && o.Scope != nil {
+	if o != nil && !IsNil(o.Scope) {
 		return true
 	}
 
@@ -379,7 +386,7 @@ func (o *IntrospectedOAuth2Token) SetScope(v string) {
 
 // GetSub returns the Sub field value if set, zero value otherwise.
 func (o *IntrospectedOAuth2Token) GetSub() string {
-	if o == nil || o.Sub == nil {
+	if o == nil || IsNil(o.Sub) {
 		var ret string
 		return ret
 	}
@@ -389,7 +396,7 @@ func (o *IntrospectedOAuth2Token) GetSub() string {
 // GetSubOk returns a tuple with the Sub field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *IntrospectedOAuth2Token) GetSubOk() (*string, bool) {
-	if o == nil || o.Sub == nil {
+	if o == nil || IsNil(o.Sub) {
 		return nil, false
 	}
 	return o.Sub, true
@@ -397,7 +404,7 @@ func (o *IntrospectedOAuth2Token) GetSubOk() (*string, bool) {
 
 // HasSub returns a boolean if a field has been set.
 func (o *IntrospectedOAuth2Token) HasSub() bool {
-	if o != nil && o.Sub != nil {
+	if o != nil && !IsNil(o.Sub) {
 		return true
 	}
 
@@ -411,7 +418,7 @@ func (o *IntrospectedOAuth2Token) SetSub(v string) {
 
 // GetTokenType returns the TokenType field value if set, zero value otherwise.
 func (o *IntrospectedOAuth2Token) GetTokenType() string {
-	if o == nil || o.TokenType == nil {
+	if o == nil || IsNil(o.TokenType) {
 		var ret string
 		return ret
 	}
@@ -421,7 +428,7 @@ func (o *IntrospectedOAuth2Token) GetTokenType() string {
 // GetTokenTypeOk returns a tuple with the TokenType field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *IntrospectedOAuth2Token) GetTokenTypeOk() (*string, bool) {
-	if o == nil || o.TokenType == nil {
+	if o == nil || IsNil(o.TokenType) {
 		return nil, false
 	}
 	return o.TokenType, true
@@ -429,7 +436,7 @@ func (o *IntrospectedOAuth2Token) GetTokenTypeOk() (*string, bool) {
 
 // HasTokenType returns a boolean if a field has been set.
 func (o *IntrospectedOAuth2Token) HasTokenType() bool {
-	if o != nil && o.TokenType != nil {
+	if o != nil && !IsNil(o.TokenType) {
 		return true
 	}
 
@@ -443,7 +450,7 @@ func (o *IntrospectedOAuth2Token) SetTokenType(v string) {
 
 // GetTokenUse returns the TokenUse field value if set, zero value otherwise.
 func (o *IntrospectedOAuth2Token) GetTokenUse() string {
-	if o == nil || o.TokenUse == nil {
+	if o == nil || IsNil(o.TokenUse) {
 		var ret string
 		return ret
 	}
@@ -453,7 +460,7 @@ func (o *IntrospectedOAuth2Token) GetTokenUse() string {
 // GetTokenUseOk returns a tuple with the TokenUse field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *IntrospectedOAuth2Token) GetTokenUseOk() (*string, bool) {
-	if o == nil || o.TokenUse == nil {
+	if o == nil || IsNil(o.TokenUse) {
 		return nil, false
 	}
 	return o.TokenUse, true
@@ -461,7 +468,7 @@ func (o *IntrospectedOAuth2Token) GetTokenUseOk() (*string, bool) {
 
 // HasTokenUse returns a boolean if a field has been set.
 func (o *IntrospectedOAuth2Token) HasTokenUse() bool {
-	if o != nil && o.TokenUse != nil {
+	if o != nil && !IsNil(o.TokenUse) {
 		return true
 	}
 
@@ -475,7 +482,7 @@ func (o *IntrospectedOAuth2Token) SetTokenUse(v string) {
 
 // GetUsername returns the Username field value if set, zero value otherwise.
 func (o *IntrospectedOAuth2Token) GetUsername() string {
-	if o == nil || o.Username == nil {
+	if o == nil || IsNil(o.Username) {
 		var ret string
 		return ret
 	}
@@ -485,7 +492,7 @@ func (o *IntrospectedOAuth2Token) GetUsername() string {
 // GetUsernameOk returns a tuple with the Username field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *IntrospectedOAuth2Token) GetUsernameOk() (*string, bool) {
-	if o == nil || o.Username == nil {
+	if o == nil || IsNil(o.Username) {
 		return nil, false
 	}
 	return o.Username, true
@@ -493,7 +500,7 @@ func (o *IntrospectedOAuth2Token) GetUsernameOk() (*string, bool) {
 
 // HasUsername returns a boolean if a field has been set.
 func (o *IntrospectedOAuth2Token) HasUsername() bool {
-	if o != nil && o.Username != nil {
+	if o != nil && !IsNil(o.Username) {
 		return true
 	}
 
@@ -506,50 +513,93 @@ func (o *IntrospectedOAuth2Token) SetUsername(v string) {
 }
 
 func (o IntrospectedOAuth2Token) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if true {
-		toSerialize["active"] = o.Active
-	}
-	if o.Aud != nil {
-		toSerialize["aud"] = o.Aud
-	}
-	if o.ClientId != nil {
-		toSerialize["client_id"] = o.ClientId
-	}
-	if o.Exp != nil {
-		toSerialize["exp"] = o.Exp
-	}
-	if o.Ext != nil {
-		toSerialize["ext"] = o.Ext
-	}
-	if o.Iat != nil {
-		toSerialize["iat"] = o.Iat
-	}
-	if o.Iss != nil {
-		toSerialize["iss"] = o.Iss
-	}
-	if o.Nbf != nil {
-		toSerialize["nbf"] = o.Nbf
-	}
-	if o.ObfuscatedSubject != nil {
-		toSerialize["obfuscated_subject"] = o.ObfuscatedSubject
-	}
-	if o.Scope != nil {
-		toSerialize["scope"] = o.Scope
-	}
-	if o.Sub != nil {
-		toSerialize["sub"] = o.Sub
-	}
-	if o.TokenType != nil {
-		toSerialize["token_type"] = o.TokenType
-	}
-	if o.TokenUse != nil {
-		toSerialize["token_use"] = o.TokenUse
-	}
-	if o.Username != nil {
-		toSerialize["username"] = o.Username
+	toSerialize, err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o IntrospectedOAuth2Token) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	toSerialize["active"] = o.Active
+	if !IsNil(o.Aud) {
+		toSerialize["aud"] = o.Aud
+	}
+	if !IsNil(o.ClientId) {
+		toSerialize["client_id"] = o.ClientId
+	}
+	if !IsNil(o.Exp) {
+		toSerialize["exp"] = o.Exp
+	}
+	if !IsNil(o.Ext) {
+		toSerialize["ext"] = o.Ext
+	}
+	if !IsNil(o.Iat) {
+		toSerialize["iat"] = o.Iat
+	}
+	if !IsNil(o.Iss) {
+		toSerialize["iss"] = o.Iss
+	}
+	if !IsNil(o.Nbf) {
+		toSerialize["nbf"] = o.Nbf
+	}
+	if !IsNil(o.ObfuscatedSubject) {
+		toSerialize["obfuscated_subject"] = o.ObfuscatedSubject
+	}
+	if !IsNil(o.Scope) {
+		toSerialize["scope"] = o.Scope
+	}
+	if !IsNil(o.Sub) {
+		toSerialize["sub"] = o.Sub
+	}
+	if !IsNil(o.TokenType) {
+		toSerialize["token_type"] = o.TokenType
+	}
+	if !IsNil(o.TokenUse) {
+		toSerialize["token_use"] = o.TokenUse
+	}
+	if !IsNil(o.Username) {
+		toSerialize["username"] = o.Username
+	}
+	return toSerialize, nil
+}
+
+func (o *IntrospectedOAuth2Token) UnmarshalJSON(data []byte) (err error) {
+	// This validates that all required properties are included in the JSON object
+	// by unmarshalling the object into a generic map with string keys and checking
+	// that every required field exists as a key in the generic map.
+	requiredProperties := []string{
+		"active",
+	}
+
+	allProperties := make(map[string]interface{})
+
+	err = json.Unmarshal(data, &allProperties)
+
+	if err != nil {
+		return err
+	}
+
+	for _, requiredProperty := range requiredProperties {
+		if _, exists := allProperties[requiredProperty]; !exists {
+			return fmt.Errorf("no value given for required property %v", requiredProperty)
+		}
+	}
+
+	varIntrospectedOAuth2Token := _IntrospectedOAuth2Token{}
+
+	decoder := json.NewDecoder(bytes.NewReader(data))
+	decoder.DisallowUnknownFields()
+	err = decoder.Decode(&varIntrospectedOAuth2Token)
+
+	if err != nil {
+		return err
+	}
+
+	*o = IntrospectedOAuth2Token(varIntrospectedOAuth2Token)
+
+	return err
 }
 
 type NullableIntrospectedOAuth2Token struct {
