@@ -15,6 +15,9 @@ import (
 	"encoding/json"
 )
 
+// checks if the IsReady200Response type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &IsReady200Response{}
+
 // IsReady200Response struct for IsReady200Response
 type IsReady200Response struct {
 	// Always \"ok\".
@@ -40,7 +43,7 @@ func NewIsReady200ResponseWithDefaults() *IsReady200Response {
 
 // GetStatus returns the Status field value if set, zero value otherwise.
 func (o *IsReady200Response) GetStatus() string {
-	if o == nil || o.Status == nil {
+	if o == nil || IsNil(o.Status) {
 		var ret string
 		return ret
 	}
@@ -50,7 +53,7 @@ func (o *IsReady200Response) GetStatus() string {
 // GetStatusOk returns a tuple with the Status field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *IsReady200Response) GetStatusOk() (*string, bool) {
-	if o == nil || o.Status == nil {
+	if o == nil || IsNil(o.Status) {
 		return nil, false
 	}
 	return o.Status, true
@@ -58,7 +61,7 @@ func (o *IsReady200Response) GetStatusOk() (*string, bool) {
 
 // HasStatus returns a boolean if a field has been set.
 func (o *IsReady200Response) HasStatus() bool {
-	if o != nil && o.Status != nil {
+	if o != nil && !IsNil(o.Status) {
 		return true
 	}
 
@@ -71,11 +74,19 @@ func (o *IsReady200Response) SetStatus(v string) {
 }
 
 func (o IsReady200Response) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.Status != nil {
-		toSerialize["status"] = o.Status
+	toSerialize, err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o IsReady200Response) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.Status) {
+		toSerialize["status"] = o.Status
+	}
+	return toSerialize, nil
 }
 
 type NullableIsReady200Response struct {
