@@ -12,8 +12,13 @@ Contact: hi@ory.sh
 package openapi
 
 import (
+	"bytes"
 	"encoding/json"
+	"fmt"
 )
+
+// checks if the JsonPatch type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &JsonPatch{}
 
 // JsonPatch A JSONPatch document as defined by RFC 6902
 type JsonPatch struct {
@@ -26,6 +31,8 @@ type JsonPatch struct {
 	// The value to be used within the operations.  Learn more [about JSON Pointers](https://datatracker.ietf.org/doc/html/rfc6901#section-5).
 	Value interface{} `json:"value,omitempty"`
 }
+
+type _JsonPatch JsonPatch
 
 // NewJsonPatch instantiates a new JsonPatch object
 // This constructor will assign default values to properties that have it defined,
@@ -48,7 +55,7 @@ func NewJsonPatchWithDefaults() *JsonPatch {
 
 // GetFrom returns the From field value if set, zero value otherwise.
 func (o *JsonPatch) GetFrom() string {
-	if o == nil || o.From == nil {
+	if o == nil || IsNil(o.From) {
 		var ret string
 		return ret
 	}
@@ -58,7 +65,7 @@ func (o *JsonPatch) GetFrom() string {
 // GetFromOk returns a tuple with the From field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *JsonPatch) GetFromOk() (*string, bool) {
-	if o == nil || o.From == nil {
+	if o == nil || IsNil(o.From) {
 		return nil, false
 	}
 	return o.From, true
@@ -66,7 +73,7 @@ func (o *JsonPatch) GetFromOk() (*string, bool) {
 
 // HasFrom returns a boolean if a field has been set.
 func (o *JsonPatch) HasFrom() bool {
-	if o != nil && o.From != nil {
+	if o != nil && !IsNil(o.From) {
 		return true
 	}
 
@@ -139,7 +146,7 @@ func (o *JsonPatch) GetValue() interface{} {
 // and a boolean to check if the value has been set.
 // NOTE: If the value is an explicit nil, `nil, true` will be returned
 func (o *JsonPatch) GetValueOk() (*interface{}, bool) {
-	if o == nil || o.Value == nil {
+	if o == nil || IsNil(o.Value) {
 		return nil, false
 	}
 	return &o.Value, true
@@ -147,7 +154,7 @@ func (o *JsonPatch) GetValueOk() (*interface{}, bool) {
 
 // HasValue returns a boolean if a field has been set.
 func (o *JsonPatch) HasValue() bool {
-	if o != nil && o.Value != nil {
+	if o != nil && IsNil(o.Value) {
 		return true
 	}
 
@@ -160,20 +167,62 @@ func (o *JsonPatch) SetValue(v interface{}) {
 }
 
 func (o JsonPatch) MarshalJSON() ([]byte, error) {
+	toSerialize, err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
+	}
+	return json.Marshal(toSerialize)
+}
+
+func (o JsonPatch) ToMap() (map[string]interface{}, error) {
 	toSerialize := map[string]interface{}{}
-	if o.From != nil {
+	if !IsNil(o.From) {
 		toSerialize["from"] = o.From
 	}
-	if true {
-		toSerialize["op"] = o.Op
-	}
-	if true {
-		toSerialize["path"] = o.Path
-	}
+	toSerialize["op"] = o.Op
+	toSerialize["path"] = o.Path
 	if o.Value != nil {
 		toSerialize["value"] = o.Value
 	}
-	return json.Marshal(toSerialize)
+	return toSerialize, nil
+}
+
+func (o *JsonPatch) UnmarshalJSON(data []byte) (err error) {
+	// This validates that all required properties are included in the JSON object
+	// by unmarshalling the object into a generic map with string keys and checking
+	// that every required field exists as a key in the generic map.
+	requiredProperties := []string{
+		"op",
+		"path",
+	}
+
+	allProperties := make(map[string]interface{})
+
+	err = json.Unmarshal(data, &allProperties)
+
+	if err != nil {
+		return err
+	}
+
+	for _, requiredProperty := range requiredProperties {
+		if _, exists := allProperties[requiredProperty]; !exists {
+			return fmt.Errorf("no value given for required property %v", requiredProperty)
+		}
+	}
+
+	varJsonPatch := _JsonPatch{}
+
+	decoder := json.NewDecoder(bytes.NewReader(data))
+	decoder.DisallowUnknownFields()
+	err = decoder.Decode(&varJsonPatch)
+
+	if err != nil {
+		return err
+	}
+
+	*o = JsonPatch(varJsonPatch)
+
+	return err
 }
 
 type NullableJsonPatch struct {
