@@ -12,8 +12,13 @@ Contact: hi@ory.sh
 package openapi
 
 import (
+	"bytes"
 	"encoding/json"
+	"fmt"
 )
+
+// checks if the JsonWebKey type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &JsonWebKey{}
 
 // JsonWebKey struct for JsonWebKey
 type JsonWebKey struct {
@@ -40,6 +45,8 @@ type JsonWebKey struct {
 	X5c []string `json:"x5c,omitempty"`
 	Y   *string  `json:"y,omitempty"`
 }
+
+type _JsonWebKey JsonWebKey
 
 // NewJsonWebKey instantiates a new JsonWebKey object
 // This constructor will assign default values to properties that have it defined,
@@ -88,7 +95,7 @@ func (o *JsonWebKey) SetAlg(v string) {
 
 // GetCrv returns the Crv field value if set, zero value otherwise.
 func (o *JsonWebKey) GetCrv() string {
-	if o == nil || o.Crv == nil {
+	if o == nil || IsNil(o.Crv) {
 		var ret string
 		return ret
 	}
@@ -98,7 +105,7 @@ func (o *JsonWebKey) GetCrv() string {
 // GetCrvOk returns a tuple with the Crv field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *JsonWebKey) GetCrvOk() (*string, bool) {
-	if o == nil || o.Crv == nil {
+	if o == nil || IsNil(o.Crv) {
 		return nil, false
 	}
 	return o.Crv, true
@@ -106,7 +113,7 @@ func (o *JsonWebKey) GetCrvOk() (*string, bool) {
 
 // HasCrv returns a boolean if a field has been set.
 func (o *JsonWebKey) HasCrv() bool {
-	if o != nil && o.Crv != nil {
+	if o != nil && !IsNil(o.Crv) {
 		return true
 	}
 
@@ -120,7 +127,7 @@ func (o *JsonWebKey) SetCrv(v string) {
 
 // GetD returns the D field value if set, zero value otherwise.
 func (o *JsonWebKey) GetD() string {
-	if o == nil || o.D == nil {
+	if o == nil || IsNil(o.D) {
 		var ret string
 		return ret
 	}
@@ -130,7 +137,7 @@ func (o *JsonWebKey) GetD() string {
 // GetDOk returns a tuple with the D field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *JsonWebKey) GetDOk() (*string, bool) {
-	if o == nil || o.D == nil {
+	if o == nil || IsNil(o.D) {
 		return nil, false
 	}
 	return o.D, true
@@ -138,7 +145,7 @@ func (o *JsonWebKey) GetDOk() (*string, bool) {
 
 // HasD returns a boolean if a field has been set.
 func (o *JsonWebKey) HasD() bool {
-	if o != nil && o.D != nil {
+	if o != nil && !IsNil(o.D) {
 		return true
 	}
 
@@ -152,7 +159,7 @@ func (o *JsonWebKey) SetD(v string) {
 
 // GetDp returns the Dp field value if set, zero value otherwise.
 func (o *JsonWebKey) GetDp() string {
-	if o == nil || o.Dp == nil {
+	if o == nil || IsNil(o.Dp) {
 		var ret string
 		return ret
 	}
@@ -162,7 +169,7 @@ func (o *JsonWebKey) GetDp() string {
 // GetDpOk returns a tuple with the Dp field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *JsonWebKey) GetDpOk() (*string, bool) {
-	if o == nil || o.Dp == nil {
+	if o == nil || IsNil(o.Dp) {
 		return nil, false
 	}
 	return o.Dp, true
@@ -170,7 +177,7 @@ func (o *JsonWebKey) GetDpOk() (*string, bool) {
 
 // HasDp returns a boolean if a field has been set.
 func (o *JsonWebKey) HasDp() bool {
-	if o != nil && o.Dp != nil {
+	if o != nil && !IsNil(o.Dp) {
 		return true
 	}
 
@@ -184,7 +191,7 @@ func (o *JsonWebKey) SetDp(v string) {
 
 // GetDq returns the Dq field value if set, zero value otherwise.
 func (o *JsonWebKey) GetDq() string {
-	if o == nil || o.Dq == nil {
+	if o == nil || IsNil(o.Dq) {
 		var ret string
 		return ret
 	}
@@ -194,7 +201,7 @@ func (o *JsonWebKey) GetDq() string {
 // GetDqOk returns a tuple with the Dq field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *JsonWebKey) GetDqOk() (*string, bool) {
-	if o == nil || o.Dq == nil {
+	if o == nil || IsNil(o.Dq) {
 		return nil, false
 	}
 	return o.Dq, true
@@ -202,7 +209,7 @@ func (o *JsonWebKey) GetDqOk() (*string, bool) {
 
 // HasDq returns a boolean if a field has been set.
 func (o *JsonWebKey) HasDq() bool {
-	if o != nil && o.Dq != nil {
+	if o != nil && !IsNil(o.Dq) {
 		return true
 	}
 
@@ -216,7 +223,7 @@ func (o *JsonWebKey) SetDq(v string) {
 
 // GetE returns the E field value if set, zero value otherwise.
 func (o *JsonWebKey) GetE() string {
-	if o == nil || o.E == nil {
+	if o == nil || IsNil(o.E) {
 		var ret string
 		return ret
 	}
@@ -226,7 +233,7 @@ func (o *JsonWebKey) GetE() string {
 // GetEOk returns a tuple with the E field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *JsonWebKey) GetEOk() (*string, bool) {
-	if o == nil || o.E == nil {
+	if o == nil || IsNil(o.E) {
 		return nil, false
 	}
 	return o.E, true
@@ -234,7 +241,7 @@ func (o *JsonWebKey) GetEOk() (*string, bool) {
 
 // HasE returns a boolean if a field has been set.
 func (o *JsonWebKey) HasE() bool {
-	if o != nil && o.E != nil {
+	if o != nil && !IsNil(o.E) {
 		return true
 	}
 
@@ -248,7 +255,7 @@ func (o *JsonWebKey) SetE(v string) {
 
 // GetK returns the K field value if set, zero value otherwise.
 func (o *JsonWebKey) GetK() string {
-	if o == nil || o.K == nil {
+	if o == nil || IsNil(o.K) {
 		var ret string
 		return ret
 	}
@@ -258,7 +265,7 @@ func (o *JsonWebKey) GetK() string {
 // GetKOk returns a tuple with the K field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *JsonWebKey) GetKOk() (*string, bool) {
-	if o == nil || o.K == nil {
+	if o == nil || IsNil(o.K) {
 		return nil, false
 	}
 	return o.K, true
@@ -266,7 +273,7 @@ func (o *JsonWebKey) GetKOk() (*string, bool) {
 
 // HasK returns a boolean if a field has been set.
 func (o *JsonWebKey) HasK() bool {
-	if o != nil && o.K != nil {
+	if o != nil && !IsNil(o.K) {
 		return true
 	}
 
@@ -328,7 +335,7 @@ func (o *JsonWebKey) SetKty(v string) {
 
 // GetN returns the N field value if set, zero value otherwise.
 func (o *JsonWebKey) GetN() string {
-	if o == nil || o.N == nil {
+	if o == nil || IsNil(o.N) {
 		var ret string
 		return ret
 	}
@@ -338,7 +345,7 @@ func (o *JsonWebKey) GetN() string {
 // GetNOk returns a tuple with the N field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *JsonWebKey) GetNOk() (*string, bool) {
-	if o == nil || o.N == nil {
+	if o == nil || IsNil(o.N) {
 		return nil, false
 	}
 	return o.N, true
@@ -346,7 +353,7 @@ func (o *JsonWebKey) GetNOk() (*string, bool) {
 
 // HasN returns a boolean if a field has been set.
 func (o *JsonWebKey) HasN() bool {
-	if o != nil && o.N != nil {
+	if o != nil && !IsNil(o.N) {
 		return true
 	}
 
@@ -360,7 +367,7 @@ func (o *JsonWebKey) SetN(v string) {
 
 // GetP returns the P field value if set, zero value otherwise.
 func (o *JsonWebKey) GetP() string {
-	if o == nil || o.P == nil {
+	if o == nil || IsNil(o.P) {
 		var ret string
 		return ret
 	}
@@ -370,7 +377,7 @@ func (o *JsonWebKey) GetP() string {
 // GetPOk returns a tuple with the P field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *JsonWebKey) GetPOk() (*string, bool) {
-	if o == nil || o.P == nil {
+	if o == nil || IsNil(o.P) {
 		return nil, false
 	}
 	return o.P, true
@@ -378,7 +385,7 @@ func (o *JsonWebKey) GetPOk() (*string, bool) {
 
 // HasP returns a boolean if a field has been set.
 func (o *JsonWebKey) HasP() bool {
-	if o != nil && o.P != nil {
+	if o != nil && !IsNil(o.P) {
 		return true
 	}
 
@@ -392,7 +399,7 @@ func (o *JsonWebKey) SetP(v string) {
 
 // GetQ returns the Q field value if set, zero value otherwise.
 func (o *JsonWebKey) GetQ() string {
-	if o == nil || o.Q == nil {
+	if o == nil || IsNil(o.Q) {
 		var ret string
 		return ret
 	}
@@ -402,7 +409,7 @@ func (o *JsonWebKey) GetQ() string {
 // GetQOk returns a tuple with the Q field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *JsonWebKey) GetQOk() (*string, bool) {
-	if o == nil || o.Q == nil {
+	if o == nil || IsNil(o.Q) {
 		return nil, false
 	}
 	return o.Q, true
@@ -410,7 +417,7 @@ func (o *JsonWebKey) GetQOk() (*string, bool) {
 
 // HasQ returns a boolean if a field has been set.
 func (o *JsonWebKey) HasQ() bool {
-	if o != nil && o.Q != nil {
+	if o != nil && !IsNil(o.Q) {
 		return true
 	}
 
@@ -424,7 +431,7 @@ func (o *JsonWebKey) SetQ(v string) {
 
 // GetQi returns the Qi field value if set, zero value otherwise.
 func (o *JsonWebKey) GetQi() string {
-	if o == nil || o.Qi == nil {
+	if o == nil || IsNil(o.Qi) {
 		var ret string
 		return ret
 	}
@@ -434,7 +441,7 @@ func (o *JsonWebKey) GetQi() string {
 // GetQiOk returns a tuple with the Qi field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *JsonWebKey) GetQiOk() (*string, bool) {
-	if o == nil || o.Qi == nil {
+	if o == nil || IsNil(o.Qi) {
 		return nil, false
 	}
 	return o.Qi, true
@@ -442,7 +449,7 @@ func (o *JsonWebKey) GetQiOk() (*string, bool) {
 
 // HasQi returns a boolean if a field has been set.
 func (o *JsonWebKey) HasQi() bool {
-	if o != nil && o.Qi != nil {
+	if o != nil && !IsNil(o.Qi) {
 		return true
 	}
 
@@ -480,7 +487,7 @@ func (o *JsonWebKey) SetUse(v string) {
 
 // GetX returns the X field value if set, zero value otherwise.
 func (o *JsonWebKey) GetX() string {
-	if o == nil || o.X == nil {
+	if o == nil || IsNil(o.X) {
 		var ret string
 		return ret
 	}
@@ -490,7 +497,7 @@ func (o *JsonWebKey) GetX() string {
 // GetXOk returns a tuple with the X field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *JsonWebKey) GetXOk() (*string, bool) {
-	if o == nil || o.X == nil {
+	if o == nil || IsNil(o.X) {
 		return nil, false
 	}
 	return o.X, true
@@ -498,7 +505,7 @@ func (o *JsonWebKey) GetXOk() (*string, bool) {
 
 // HasX returns a boolean if a field has been set.
 func (o *JsonWebKey) HasX() bool {
-	if o != nil && o.X != nil {
+	if o != nil && !IsNil(o.X) {
 		return true
 	}
 
@@ -512,7 +519,7 @@ func (o *JsonWebKey) SetX(v string) {
 
 // GetX5c returns the X5c field value if set, zero value otherwise.
 func (o *JsonWebKey) GetX5c() []string {
-	if o == nil || o.X5c == nil {
+	if o == nil || IsNil(o.X5c) {
 		var ret []string
 		return ret
 	}
@@ -522,7 +529,7 @@ func (o *JsonWebKey) GetX5c() []string {
 // GetX5cOk returns a tuple with the X5c field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *JsonWebKey) GetX5cOk() ([]string, bool) {
-	if o == nil || o.X5c == nil {
+	if o == nil || IsNil(o.X5c) {
 		return nil, false
 	}
 	return o.X5c, true
@@ -530,7 +537,7 @@ func (o *JsonWebKey) GetX5cOk() ([]string, bool) {
 
 // HasX5c returns a boolean if a field has been set.
 func (o *JsonWebKey) HasX5c() bool {
-	if o != nil && o.X5c != nil {
+	if o != nil && !IsNil(o.X5c) {
 		return true
 	}
 
@@ -544,7 +551,7 @@ func (o *JsonWebKey) SetX5c(v []string) {
 
 // GetY returns the Y field value if set, zero value otherwise.
 func (o *JsonWebKey) GetY() string {
-	if o == nil || o.Y == nil {
+	if o == nil || IsNil(o.Y) {
 		var ret string
 		return ret
 	}
@@ -554,7 +561,7 @@ func (o *JsonWebKey) GetY() string {
 // GetYOk returns a tuple with the Y field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *JsonWebKey) GetYOk() (*string, bool) {
-	if o == nil || o.Y == nil {
+	if o == nil || IsNil(o.Y) {
 		return nil, false
 	}
 	return o.Y, true
@@ -562,7 +569,7 @@ func (o *JsonWebKey) GetYOk() (*string, bool) {
 
 // HasY returns a boolean if a field has been set.
 func (o *JsonWebKey) HasY() bool {
-	if o != nil && o.Y != nil {
+	if o != nil && !IsNil(o.Y) {
 		return true
 	}
 
@@ -575,59 +582,99 @@ func (o *JsonWebKey) SetY(v string) {
 }
 
 func (o JsonWebKey) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if true {
-		toSerialize["alg"] = o.Alg
-	}
-	if o.Crv != nil {
-		toSerialize["crv"] = o.Crv
-	}
-	if o.D != nil {
-		toSerialize["d"] = o.D
-	}
-	if o.Dp != nil {
-		toSerialize["dp"] = o.Dp
-	}
-	if o.Dq != nil {
-		toSerialize["dq"] = o.Dq
-	}
-	if o.E != nil {
-		toSerialize["e"] = o.E
-	}
-	if o.K != nil {
-		toSerialize["k"] = o.K
-	}
-	if true {
-		toSerialize["kid"] = o.Kid
-	}
-	if true {
-		toSerialize["kty"] = o.Kty
-	}
-	if o.N != nil {
-		toSerialize["n"] = o.N
-	}
-	if o.P != nil {
-		toSerialize["p"] = o.P
-	}
-	if o.Q != nil {
-		toSerialize["q"] = o.Q
-	}
-	if o.Qi != nil {
-		toSerialize["qi"] = o.Qi
-	}
-	if true {
-		toSerialize["use"] = o.Use
-	}
-	if o.X != nil {
-		toSerialize["x"] = o.X
-	}
-	if o.X5c != nil {
-		toSerialize["x5c"] = o.X5c
-	}
-	if o.Y != nil {
-		toSerialize["y"] = o.Y
+	toSerialize, err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o JsonWebKey) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	toSerialize["alg"] = o.Alg
+	if !IsNil(o.Crv) {
+		toSerialize["crv"] = o.Crv
+	}
+	if !IsNil(o.D) {
+		toSerialize["d"] = o.D
+	}
+	if !IsNil(o.Dp) {
+		toSerialize["dp"] = o.Dp
+	}
+	if !IsNil(o.Dq) {
+		toSerialize["dq"] = o.Dq
+	}
+	if !IsNil(o.E) {
+		toSerialize["e"] = o.E
+	}
+	if !IsNil(o.K) {
+		toSerialize["k"] = o.K
+	}
+	toSerialize["kid"] = o.Kid
+	toSerialize["kty"] = o.Kty
+	if !IsNil(o.N) {
+		toSerialize["n"] = o.N
+	}
+	if !IsNil(o.P) {
+		toSerialize["p"] = o.P
+	}
+	if !IsNil(o.Q) {
+		toSerialize["q"] = o.Q
+	}
+	if !IsNil(o.Qi) {
+		toSerialize["qi"] = o.Qi
+	}
+	toSerialize["use"] = o.Use
+	if !IsNil(o.X) {
+		toSerialize["x"] = o.X
+	}
+	if !IsNil(o.X5c) {
+		toSerialize["x5c"] = o.X5c
+	}
+	if !IsNil(o.Y) {
+		toSerialize["y"] = o.Y
+	}
+	return toSerialize, nil
+}
+
+func (o *JsonWebKey) UnmarshalJSON(data []byte) (err error) {
+	// This validates that all required properties are included in the JSON object
+	// by unmarshalling the object into a generic map with string keys and checking
+	// that every required field exists as a key in the generic map.
+	requiredProperties := []string{
+		"alg",
+		"kid",
+		"kty",
+		"use",
+	}
+
+	allProperties := make(map[string]interface{})
+
+	err = json.Unmarshal(data, &allProperties)
+
+	if err != nil {
+		return err
+	}
+
+	for _, requiredProperty := range requiredProperties {
+		if _, exists := allProperties[requiredProperty]; !exists {
+			return fmt.Errorf("no value given for required property %v", requiredProperty)
+		}
+	}
+
+	varJsonWebKey := _JsonWebKey{}
+
+	decoder := json.NewDecoder(bytes.NewReader(data))
+	decoder.DisallowUnknownFields()
+	err = decoder.Decode(&varJsonWebKey)
+
+	if err != nil {
+		return err
+	}
+
+	*o = JsonWebKey(varJsonWebKey)
+
+	return err
 }
 
 type NullableJsonWebKey struct {
