@@ -15,6 +15,9 @@ import (
 	"encoding/json"
 )
 
+// checks if the JsonWebKeySet type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &JsonWebKeySet{}
+
 // JsonWebKeySet JSON Web Key Set
 type JsonWebKeySet struct {
 	// List of JSON Web Keys  The value of the \"keys\" parameter is an array of JSON Web Key (JWK) values. By default, the order of the JWK values within the array does not imply an order of preference among them, although applications of JWK Sets can choose to assign a meaning to the order for their purposes, if desired.
@@ -40,7 +43,7 @@ func NewJsonWebKeySetWithDefaults() *JsonWebKeySet {
 
 // GetKeys returns the Keys field value if set, zero value otherwise.
 func (o *JsonWebKeySet) GetKeys() []JsonWebKey {
-	if o == nil || o.Keys == nil {
+	if o == nil || IsNil(o.Keys) {
 		var ret []JsonWebKey
 		return ret
 	}
@@ -50,7 +53,7 @@ func (o *JsonWebKeySet) GetKeys() []JsonWebKey {
 // GetKeysOk returns a tuple with the Keys field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *JsonWebKeySet) GetKeysOk() ([]JsonWebKey, bool) {
-	if o == nil || o.Keys == nil {
+	if o == nil || IsNil(o.Keys) {
 		return nil, false
 	}
 	return o.Keys, true
@@ -58,7 +61,7 @@ func (o *JsonWebKeySet) GetKeysOk() ([]JsonWebKey, bool) {
 
 // HasKeys returns a boolean if a field has been set.
 func (o *JsonWebKeySet) HasKeys() bool {
-	if o != nil && o.Keys != nil {
+	if o != nil && !IsNil(o.Keys) {
 		return true
 	}
 
@@ -71,11 +74,19 @@ func (o *JsonWebKeySet) SetKeys(v []JsonWebKey) {
 }
 
 func (o JsonWebKeySet) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.Keys != nil {
-		toSerialize["keys"] = o.Keys
+	toSerialize, err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o JsonWebKeySet) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.Keys) {
+		toSerialize["keys"] = o.Keys
+	}
+	return toSerialize, nil
 }
 
 type NullableJsonWebKeySet struct {
