@@ -12,8 +12,13 @@ Contact: hi@ory.sh
 package openapi
 
 import (
+	"bytes"
 	"encoding/json"
+	"fmt"
 )
+
+// checks if the OAuth2ConsentRequest type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &OAuth2ConsentRequest{}
 
 // OAuth2ConsentRequest struct for OAuth2ConsentRequest
 type OAuth2ConsentRequest struct {
@@ -39,6 +44,8 @@ type OAuth2ConsentRequest struct {
 	Subject *string `json:"subject,omitempty"`
 }
 
+type _OAuth2ConsentRequest OAuth2ConsentRequest
+
 // NewOAuth2ConsentRequest instantiates a new OAuth2ConsentRequest object
 // This constructor will assign default values to properties that have it defined,
 // and makes sure properties required by API are set, but the set of arguments
@@ -59,7 +66,7 @@ func NewOAuth2ConsentRequestWithDefaults() *OAuth2ConsentRequest {
 
 // GetAcr returns the Acr field value if set, zero value otherwise.
 func (o *OAuth2ConsentRequest) GetAcr() string {
-	if o == nil || o.Acr == nil {
+	if o == nil || IsNil(o.Acr) {
 		var ret string
 		return ret
 	}
@@ -69,7 +76,7 @@ func (o *OAuth2ConsentRequest) GetAcr() string {
 // GetAcrOk returns a tuple with the Acr field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2ConsentRequest) GetAcrOk() (*string, bool) {
-	if o == nil || o.Acr == nil {
+	if o == nil || IsNil(o.Acr) {
 		return nil, false
 	}
 	return o.Acr, true
@@ -77,7 +84,7 @@ func (o *OAuth2ConsentRequest) GetAcrOk() (*string, bool) {
 
 // HasAcr returns a boolean if a field has been set.
 func (o *OAuth2ConsentRequest) HasAcr() bool {
-	if o != nil && o.Acr != nil {
+	if o != nil && !IsNil(o.Acr) {
 		return true
 	}
 
@@ -91,7 +98,7 @@ func (o *OAuth2ConsentRequest) SetAcr(v string) {
 
 // GetAmr returns the Amr field value if set, zero value otherwise.
 func (o *OAuth2ConsentRequest) GetAmr() []string {
-	if o == nil || o.Amr == nil {
+	if o == nil || IsNil(o.Amr) {
 		var ret []string
 		return ret
 	}
@@ -101,7 +108,7 @@ func (o *OAuth2ConsentRequest) GetAmr() []string {
 // GetAmrOk returns a tuple with the Amr field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2ConsentRequest) GetAmrOk() ([]string, bool) {
-	if o == nil || o.Amr == nil {
+	if o == nil || IsNil(o.Amr) {
 		return nil, false
 	}
 	return o.Amr, true
@@ -109,7 +116,7 @@ func (o *OAuth2ConsentRequest) GetAmrOk() ([]string, bool) {
 
 // HasAmr returns a boolean if a field has been set.
 func (o *OAuth2ConsentRequest) HasAmr() bool {
-	if o != nil && o.Amr != nil {
+	if o != nil && !IsNil(o.Amr) {
 		return true
 	}
 
@@ -147,7 +154,7 @@ func (o *OAuth2ConsentRequest) SetChallenge(v string) {
 
 // GetClient returns the Client field value if set, zero value otherwise.
 func (o *OAuth2ConsentRequest) GetClient() OAuth2Client {
-	if o == nil || o.Client == nil {
+	if o == nil || IsNil(o.Client) {
 		var ret OAuth2Client
 		return ret
 	}
@@ -157,7 +164,7 @@ func (o *OAuth2ConsentRequest) GetClient() OAuth2Client {
 // GetClientOk returns a tuple with the Client field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2ConsentRequest) GetClientOk() (*OAuth2Client, bool) {
-	if o == nil || o.Client == nil {
+	if o == nil || IsNil(o.Client) {
 		return nil, false
 	}
 	return o.Client, true
@@ -165,7 +172,7 @@ func (o *OAuth2ConsentRequest) GetClientOk() (*OAuth2Client, bool) {
 
 // HasClient returns a boolean if a field has been set.
 func (o *OAuth2ConsentRequest) HasClient() bool {
-	if o != nil && o.Client != nil {
+	if o != nil && !IsNil(o.Client) {
 		return true
 	}
 
@@ -190,7 +197,7 @@ func (o *OAuth2ConsentRequest) GetContext() interface{} {
 // and a boolean to check if the value has been set.
 // NOTE: If the value is an explicit nil, `nil, true` will be returned
 func (o *OAuth2ConsentRequest) GetContextOk() (*interface{}, bool) {
-	if o == nil || o.Context == nil {
+	if o == nil || IsNil(o.Context) {
 		return nil, false
 	}
 	return &o.Context, true
@@ -198,7 +205,7 @@ func (o *OAuth2ConsentRequest) GetContextOk() (*interface{}, bool) {
 
 // HasContext returns a boolean if a field has been set.
 func (o *OAuth2ConsentRequest) HasContext() bool {
-	if o != nil && o.Context != nil {
+	if o != nil && IsNil(o.Context) {
 		return true
 	}
 
@@ -212,7 +219,7 @@ func (o *OAuth2ConsentRequest) SetContext(v interface{}) {
 
 // GetLoginChallenge returns the LoginChallenge field value if set, zero value otherwise.
 func (o *OAuth2ConsentRequest) GetLoginChallenge() string {
-	if o == nil || o.LoginChallenge == nil {
+	if o == nil || IsNil(o.LoginChallenge) {
 		var ret string
 		return ret
 	}
@@ -222,7 +229,7 @@ func (o *OAuth2ConsentRequest) GetLoginChallenge() string {
 // GetLoginChallengeOk returns a tuple with the LoginChallenge field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2ConsentRequest) GetLoginChallengeOk() (*string, bool) {
-	if o == nil || o.LoginChallenge == nil {
+	if o == nil || IsNil(o.LoginChallenge) {
 		return nil, false
 	}
 	return o.LoginChallenge, true
@@ -230,7 +237,7 @@ func (o *OAuth2ConsentRequest) GetLoginChallengeOk() (*string, bool) {
 
 // HasLoginChallenge returns a boolean if a field has been set.
 func (o *OAuth2ConsentRequest) HasLoginChallenge() bool {
-	if o != nil && o.LoginChallenge != nil {
+	if o != nil && !IsNil(o.LoginChallenge) {
 		return true
 	}
 
@@ -244,7 +251,7 @@ func (o *OAuth2ConsentRequest) SetLoginChallenge(v string) {
 
 // GetLoginSessionId returns the LoginSessionId field value if set, zero value otherwise.
 func (o *OAuth2ConsentRequest) GetLoginSessionId() string {
-	if o == nil || o.LoginSessionId == nil {
+	if o == nil || IsNil(o.LoginSessionId) {
 		var ret string
 		return ret
 	}
@@ -254,7 +261,7 @@ func (o *OAuth2ConsentRequest) GetLoginSessionId() string {
 // GetLoginSessionIdOk returns a tuple with the LoginSessionId field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2ConsentRequest) GetLoginSessionIdOk() (*string, bool) {
-	if o == nil || o.LoginSessionId == nil {
+	if o == nil || IsNil(o.LoginSessionId) {
 		return nil, false
 	}
 	return o.LoginSessionId, true
@@ -262,7 +269,7 @@ func (o *OAuth2ConsentRequest) GetLoginSessionIdOk() (*string, bool) {
 
 // HasLoginSessionId returns a boolean if a field has been set.
 func (o *OAuth2ConsentRequest) HasLoginSessionId() bool {
-	if o != nil && o.LoginSessionId != nil {
+	if o != nil && !IsNil(o.LoginSessionId) {
 		return true
 	}
 
@@ -276,7 +283,7 @@ func (o *OAuth2ConsentRequest) SetLoginSessionId(v string) {
 
 // GetOidcContext returns the OidcContext field value if set, zero value otherwise.
 func (o *OAuth2ConsentRequest) GetOidcContext() OAuth2ConsentRequestOpenIDConnectContext {
-	if o == nil || o.OidcContext == nil {
+	if o == nil || IsNil(o.OidcContext) {
 		var ret OAuth2ConsentRequestOpenIDConnectContext
 		return ret
 	}
@@ -286,7 +293,7 @@ func (o *OAuth2ConsentRequest) GetOidcContext() OAuth2ConsentRequestOpenIDConnec
 // GetOidcContextOk returns a tuple with the OidcContext field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2ConsentRequest) GetOidcContextOk() (*OAuth2ConsentRequestOpenIDConnectContext, bool) {
-	if o == nil || o.OidcContext == nil {
+	if o == nil || IsNil(o.OidcContext) {
 		return nil, false
 	}
 	return o.OidcContext, true
@@ -294,7 +301,7 @@ func (o *OAuth2ConsentRequest) GetOidcContextOk() (*OAuth2ConsentRequestOpenIDCo
 
 // HasOidcContext returns a boolean if a field has been set.
 func (o *OAuth2ConsentRequest) HasOidcContext() bool {
-	if o != nil && o.OidcContext != nil {
+	if o != nil && !IsNil(o.OidcContext) {
 		return true
 	}
 
@@ -308,7 +315,7 @@ func (o *OAuth2ConsentRequest) SetOidcContext(v OAuth2ConsentRequestOpenIDConnec
 
 // GetRequestUrl returns the RequestUrl field value if set, zero value otherwise.
 func (o *OAuth2ConsentRequest) GetRequestUrl() string {
-	if o == nil || o.RequestUrl == nil {
+	if o == nil || IsNil(o.RequestUrl) {
 		var ret string
 		return ret
 	}
@@ -318,7 +325,7 @@ func (o *OAuth2ConsentRequest) GetRequestUrl() string {
 // GetRequestUrlOk returns a tuple with the RequestUrl field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2ConsentRequest) GetRequestUrlOk() (*string, bool) {
-	if o == nil || o.RequestUrl == nil {
+	if o == nil || IsNil(o.RequestUrl) {
 		return nil, false
 	}
 	return o.RequestUrl, true
@@ -326,7 +333,7 @@ func (o *OAuth2ConsentRequest) GetRequestUrlOk() (*string, bool) {
 
 // HasRequestUrl returns a boolean if a field has been set.
 func (o *OAuth2ConsentRequest) HasRequestUrl() bool {
-	if o != nil && o.RequestUrl != nil {
+	if o != nil && !IsNil(o.RequestUrl) {
 		return true
 	}
 
@@ -340,7 +347,7 @@ func (o *OAuth2ConsentRequest) SetRequestUrl(v string) {
 
 // GetRequestedAccessTokenAudience returns the RequestedAccessTokenAudience field value if set, zero value otherwise.
 func (o *OAuth2ConsentRequest) GetRequestedAccessTokenAudience() []string {
-	if o == nil || o.RequestedAccessTokenAudience == nil {
+	if o == nil || IsNil(o.RequestedAccessTokenAudience) {
 		var ret []string
 		return ret
 	}
@@ -350,7 +357,7 @@ func (o *OAuth2ConsentRequest) GetRequestedAccessTokenAudience() []string {
 // GetRequestedAccessTokenAudienceOk returns a tuple with the RequestedAccessTokenAudience field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2ConsentRequest) GetRequestedAccessTokenAudienceOk() ([]string, bool) {
-	if o == nil || o.RequestedAccessTokenAudience == nil {
+	if o == nil || IsNil(o.RequestedAccessTokenAudience) {
 		return nil, false
 	}
 	return o.RequestedAccessTokenAudience, true
@@ -358,7 +365,7 @@ func (o *OAuth2ConsentRequest) GetRequestedAccessTokenAudienceOk() ([]string, bo
 
 // HasRequestedAccessTokenAudience returns a boolean if a field has been set.
 func (o *OAuth2ConsentRequest) HasRequestedAccessTokenAudience() bool {
-	if o != nil && o.RequestedAccessTokenAudience != nil {
+	if o != nil && !IsNil(o.RequestedAccessTokenAudience) {
 		return true
 	}
 
@@ -372,7 +379,7 @@ func (o *OAuth2ConsentRequest) SetRequestedAccessTokenAudience(v []string) {
 
 // GetRequestedScope returns the RequestedScope field value if set, zero value otherwise.
 func (o *OAuth2ConsentRequest) GetRequestedScope() []string {
-	if o == nil || o.RequestedScope == nil {
+	if o == nil || IsNil(o.RequestedScope) {
 		var ret []string
 		return ret
 	}
@@ -382,7 +389,7 @@ func (o *OAuth2ConsentRequest) GetRequestedScope() []string {
 // GetRequestedScopeOk returns a tuple with the RequestedScope field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2ConsentRequest) GetRequestedScopeOk() ([]string, bool) {
-	if o == nil || o.RequestedScope == nil {
+	if o == nil || IsNil(o.RequestedScope) {
 		return nil, false
 	}
 	return o.RequestedScope, true
@@ -390,7 +397,7 @@ func (o *OAuth2ConsentRequest) GetRequestedScopeOk() ([]string, bool) {
 
 // HasRequestedScope returns a boolean if a field has been set.
 func (o *OAuth2ConsentRequest) HasRequestedScope() bool {
-	if o != nil && o.RequestedScope != nil {
+	if o != nil && !IsNil(o.RequestedScope) {
 		return true
 	}
 
@@ -404,7 +411,7 @@ func (o *OAuth2ConsentRequest) SetRequestedScope(v []string) {
 
 // GetSkip returns the Skip field value if set, zero value otherwise.
 func (o *OAuth2ConsentRequest) GetSkip() bool {
-	if o == nil || o.Skip == nil {
+	if o == nil || IsNil(o.Skip) {
 		var ret bool
 		return ret
 	}
@@ -414,7 +421,7 @@ func (o *OAuth2ConsentRequest) GetSkip() bool {
 // GetSkipOk returns a tuple with the Skip field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2ConsentRequest) GetSkipOk() (*bool, bool) {
-	if o == nil || o.Skip == nil {
+	if o == nil || IsNil(o.Skip) {
 		return nil, false
 	}
 	return o.Skip, true
@@ -422,7 +429,7 @@ func (o *OAuth2ConsentRequest) GetSkipOk() (*bool, bool) {
 
 // HasSkip returns a boolean if a field has been set.
 func (o *OAuth2ConsentRequest) HasSkip() bool {
-	if o != nil && o.Skip != nil {
+	if o != nil && !IsNil(o.Skip) {
 		return true
 	}
 
@@ -436,7 +443,7 @@ func (o *OAuth2ConsentRequest) SetSkip(v bool) {
 
 // GetSubject returns the Subject field value if set, zero value otherwise.
 func (o *OAuth2ConsentRequest) GetSubject() string {
-	if o == nil || o.Subject == nil {
+	if o == nil || IsNil(o.Subject) {
 		var ret string
 		return ret
 	}
@@ -446,7 +453,7 @@ func (o *OAuth2ConsentRequest) GetSubject() string {
 // GetSubjectOk returns a tuple with the Subject field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2ConsentRequest) GetSubjectOk() (*string, bool) {
-	if o == nil || o.Subject == nil {
+	if o == nil || IsNil(o.Subject) {
 		return nil, false
 	}
 	return o.Subject, true
@@ -454,7 +461,7 @@ func (o *OAuth2ConsentRequest) GetSubjectOk() (*string, bool) {
 
 // HasSubject returns a boolean if a field has been set.
 func (o *OAuth2ConsentRequest) HasSubject() bool {
-	if o != nil && o.Subject != nil {
+	if o != nil && !IsNil(o.Subject) {
 		return true
 	}
 
@@ -467,47 +474,90 @@ func (o *OAuth2ConsentRequest) SetSubject(v string) {
 }
 
 func (o OAuth2ConsentRequest) MarshalJSON() ([]byte, error) {
+	toSerialize, err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
+	}
+	return json.Marshal(toSerialize)
+}
+
+func (o OAuth2ConsentRequest) ToMap() (map[string]interface{}, error) {
 	toSerialize := map[string]interface{}{}
-	if o.Acr != nil {
+	if !IsNil(o.Acr) {
 		toSerialize["acr"] = o.Acr
 	}
-	if o.Amr != nil {
+	if !IsNil(o.Amr) {
 		toSerialize["amr"] = o.Amr
 	}
-	if true {
-		toSerialize["challenge"] = o.Challenge
-	}
-	if o.Client != nil {
+	toSerialize["challenge"] = o.Challenge
+	if !IsNil(o.Client) {
 		toSerialize["client"] = o.Client
 	}
 	if o.Context != nil {
 		toSerialize["context"] = o.Context
 	}
-	if o.LoginChallenge != nil {
+	if !IsNil(o.LoginChallenge) {
 		toSerialize["login_challenge"] = o.LoginChallenge
 	}
-	if o.LoginSessionId != nil {
+	if !IsNil(o.LoginSessionId) {
 		toSerialize["login_session_id"] = o.LoginSessionId
 	}
-	if o.OidcContext != nil {
+	if !IsNil(o.OidcContext) {
 		toSerialize["oidc_context"] = o.OidcContext
 	}
-	if o.RequestUrl != nil {
+	if !IsNil(o.RequestUrl) {
 		toSerialize["request_url"] = o.RequestUrl
 	}
-	if o.RequestedAccessTokenAudience != nil {
+	if !IsNil(o.RequestedAccessTokenAudience) {
 		toSerialize["requested_access_token_audience"] = o.RequestedAccessTokenAudience
 	}
-	if o.RequestedScope != nil {
+	if !IsNil(o.RequestedScope) {
 		toSerialize["requested_scope"] = o.RequestedScope
 	}
-	if o.Skip != nil {
+	if !IsNil(o.Skip) {
 		toSerialize["skip"] = o.Skip
 	}
-	if o.Subject != nil {
+	if !IsNil(o.Subject) {
 		toSerialize["subject"] = o.Subject
 	}
-	return json.Marshal(toSerialize)
+	return toSerialize, nil
+}
+
+func (o *OAuth2ConsentRequest) UnmarshalJSON(data []byte) (err error) {
+	// This validates that all required properties are included in the JSON object
+	// by unmarshalling the object into a generic map with string keys and checking
+	// that every required field exists as a key in the generic map.
+	requiredProperties := []string{
+		"challenge",
+	}
+
+	allProperties := make(map[string]interface{})
+
+	err = json.Unmarshal(data, &allProperties)
+
+	if err != nil {
+		return err
+	}
+
+	for _, requiredProperty := range requiredProperties {
+		if _, exists := allProperties[requiredProperty]; !exists {
+			return fmt.Errorf("no value given for required property %v", requiredProperty)
+		}
+	}
+
+	varOAuth2ConsentRequest := _OAuth2ConsentRequest{}
+
+	decoder := json.NewDecoder(bytes.NewReader(data))
+	decoder.DisallowUnknownFields()
+	err = decoder.Decode(&varOAuth2ConsentRequest)
+
+	if err != nil {
+		return err
+	}
+
+	*o = OAuth2ConsentRequest(varOAuth2ConsentRequest)
+
+	return err
 }
 
 type NullableOAuth2ConsentRequest struct {
