@@ -16,6 +16,9 @@ import (
 	"time"
 )
 
+// checks if the OAuth2ConsentSessionExpiresAt type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &OAuth2ConsentSessionExpiresAt{}
+
 // OAuth2ConsentSessionExpiresAt struct for OAuth2ConsentSessionExpiresAt
 type OAuth2ConsentSessionExpiresAt struct {
 	AccessToken   *time.Time `json:"access_token,omitempty"`
@@ -44,7 +47,7 @@ func NewOAuth2ConsentSessionExpiresAtWithDefaults() *OAuth2ConsentSessionExpires
 
 // GetAccessToken returns the AccessToken field value if set, zero value otherwise.
 func (o *OAuth2ConsentSessionExpiresAt) GetAccessToken() time.Time {
-	if o == nil || o.AccessToken == nil {
+	if o == nil || IsNil(o.AccessToken) {
 		var ret time.Time
 		return ret
 	}
@@ -54,7 +57,7 @@ func (o *OAuth2ConsentSessionExpiresAt) GetAccessToken() time.Time {
 // GetAccessTokenOk returns a tuple with the AccessToken field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2ConsentSessionExpiresAt) GetAccessTokenOk() (*time.Time, bool) {
-	if o == nil || o.AccessToken == nil {
+	if o == nil || IsNil(o.AccessToken) {
 		return nil, false
 	}
 	return o.AccessToken, true
@@ -62,7 +65,7 @@ func (o *OAuth2ConsentSessionExpiresAt) GetAccessTokenOk() (*time.Time, bool) {
 
 // HasAccessToken returns a boolean if a field has been set.
 func (o *OAuth2ConsentSessionExpiresAt) HasAccessToken() bool {
-	if o != nil && o.AccessToken != nil {
+	if o != nil && !IsNil(o.AccessToken) {
 		return true
 	}
 
@@ -76,7 +79,7 @@ func (o *OAuth2ConsentSessionExpiresAt) SetAccessToken(v time.Time) {
 
 // GetAuthorizeCode returns the AuthorizeCode field value if set, zero value otherwise.
 func (o *OAuth2ConsentSessionExpiresAt) GetAuthorizeCode() time.Time {
-	if o == nil || o.AuthorizeCode == nil {
+	if o == nil || IsNil(o.AuthorizeCode) {
 		var ret time.Time
 		return ret
 	}
@@ -86,7 +89,7 @@ func (o *OAuth2ConsentSessionExpiresAt) GetAuthorizeCode() time.Time {
 // GetAuthorizeCodeOk returns a tuple with the AuthorizeCode field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2ConsentSessionExpiresAt) GetAuthorizeCodeOk() (*time.Time, bool) {
-	if o == nil || o.AuthorizeCode == nil {
+	if o == nil || IsNil(o.AuthorizeCode) {
 		return nil, false
 	}
 	return o.AuthorizeCode, true
@@ -94,7 +97,7 @@ func (o *OAuth2ConsentSessionExpiresAt) GetAuthorizeCodeOk() (*time.Time, bool) 
 
 // HasAuthorizeCode returns a boolean if a field has been set.
 func (o *OAuth2ConsentSessionExpiresAt) HasAuthorizeCode() bool {
-	if o != nil && o.AuthorizeCode != nil {
+	if o != nil && !IsNil(o.AuthorizeCode) {
 		return true
 	}
 
@@ -108,7 +111,7 @@ func (o *OAuth2ConsentSessionExpiresAt) SetAuthorizeCode(v time.Time) {
 
 // GetIdToken returns the IdToken field value if set, zero value otherwise.
 func (o *OAuth2ConsentSessionExpiresAt) GetIdToken() time.Time {
-	if o == nil || o.IdToken == nil {
+	if o == nil || IsNil(o.IdToken) {
 		var ret time.Time
 		return ret
 	}
@@ -118,7 +121,7 @@ func (o *OAuth2ConsentSessionExpiresAt) GetIdToken() time.Time {
 // GetIdTokenOk returns a tuple with the IdToken field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2ConsentSessionExpiresAt) GetIdTokenOk() (*time.Time, bool) {
-	if o == nil || o.IdToken == nil {
+	if o == nil || IsNil(o.IdToken) {
 		return nil, false
 	}
 	return o.IdToken, true
@@ -126,7 +129,7 @@ func (o *OAuth2ConsentSessionExpiresAt) GetIdTokenOk() (*time.Time, bool) {
 
 // HasIdToken returns a boolean if a field has been set.
 func (o *OAuth2ConsentSessionExpiresAt) HasIdToken() bool {
-	if o != nil && o.IdToken != nil {
+	if o != nil && !IsNil(o.IdToken) {
 		return true
 	}
 
@@ -140,7 +143,7 @@ func (o *OAuth2ConsentSessionExpiresAt) SetIdToken(v time.Time) {
 
 // GetParContext returns the ParContext field value if set, zero value otherwise.
 func (o *OAuth2ConsentSessionExpiresAt) GetParContext() time.Time {
-	if o == nil || o.ParContext == nil {
+	if o == nil || IsNil(o.ParContext) {
 		var ret time.Time
 		return ret
 	}
@@ -150,7 +153,7 @@ func (o *OAuth2ConsentSessionExpiresAt) GetParContext() time.Time {
 // GetParContextOk returns a tuple with the ParContext field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2ConsentSessionExpiresAt) GetParContextOk() (*time.Time, bool) {
-	if o == nil || o.ParContext == nil {
+	if o == nil || IsNil(o.ParContext) {
 		return nil, false
 	}
 	return o.ParContext, true
@@ -158,7 +161,7 @@ func (o *OAuth2ConsentSessionExpiresAt) GetParContextOk() (*time.Time, bool) {
 
 // HasParContext returns a boolean if a field has been set.
 func (o *OAuth2ConsentSessionExpiresAt) HasParContext() bool {
-	if o != nil && o.ParContext != nil {
+	if o != nil && !IsNil(o.ParContext) {
 		return true
 	}
 
@@ -172,7 +175,7 @@ func (o *OAuth2ConsentSessionExpiresAt) SetParContext(v time.Time) {
 
 // GetRefreshToken returns the RefreshToken field value if set, zero value otherwise.
 func (o *OAuth2ConsentSessionExpiresAt) GetRefreshToken() time.Time {
-	if o == nil || o.RefreshToken == nil {
+	if o == nil || IsNil(o.RefreshToken) {
 		var ret time.Time
 		return ret
 	}
@@ -182,7 +185,7 @@ func (o *OAuth2ConsentSessionExpiresAt) GetRefreshToken() time.Time {
 // GetRefreshTokenOk returns a tuple with the RefreshToken field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2ConsentSessionExpiresAt) GetRefreshTokenOk() (*time.Time, bool) {
-	if o == nil || o.RefreshToken == nil {
+	if o == nil || IsNil(o.RefreshToken) {
 		return nil, false
 	}
 	return o.RefreshToken, true
@@ -190,7 +193,7 @@ func (o *OAuth2ConsentSessionExpiresAt) GetRefreshTokenOk() (*time.Time, bool) {
 
 // HasRefreshToken returns a boolean if a field has been set.
 func (o *OAuth2ConsentSessionExpiresAt) HasRefreshToken() bool {
-	if o != nil && o.RefreshToken != nil {
+	if o != nil && !IsNil(o.RefreshToken) {
 		return true
 	}
 
@@ -203,23 +206,31 @@ func (o *OAuth2ConsentSessionExpiresAt) SetRefreshToken(v time.Time) {
 }
 
 func (o OAuth2ConsentSessionExpiresAt) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.AccessToken != nil {
-		toSerialize["access_token"] = o.AccessToken
-	}
-	if o.AuthorizeCode != nil {
-		toSerialize["authorize_code"] = o.AuthorizeCode
-	}
-	if o.IdToken != nil {
-		toSerialize["id_token"] = o.IdToken
-	}
-	if o.ParContext != nil {
-		toSerialize["par_context"] = o.ParContext
-	}
-	if o.RefreshToken != nil {
-		toSerialize["refresh_token"] = o.RefreshToken
+	toSerialize, err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o OAuth2ConsentSessionExpiresAt) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.AccessToken) {
+		toSerialize["access_token"] = o.AccessToken
+	}
+	if !IsNil(o.AuthorizeCode) {
+		toSerialize["authorize_code"] = o.AuthorizeCode
+	}
+	if !IsNil(o.IdToken) {
+		toSerialize["id_token"] = o.IdToken
+	}
+	if !IsNil(o.ParContext) {
+		toSerialize["par_context"] = o.ParContext
+	}
+	if !IsNil(o.RefreshToken) {
+		toSerialize["refresh_token"] = o.RefreshToken
+	}
+	return toSerialize, nil
 }
 
 type NullableOAuth2ConsentSessionExpiresAt struct {
