@@ -15,6 +15,9 @@ import (
 	"encoding/json"
 )
 
+// checks if the OAuth2TokenExchange type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &OAuth2TokenExchange{}
+
 // OAuth2TokenExchange OAuth2 Token Exchange Result
 type OAuth2TokenExchange struct {
 	// The access token issued by the authorization server.
@@ -50,7 +53,7 @@ func NewOAuth2TokenExchangeWithDefaults() *OAuth2TokenExchange {
 
 // GetAccessToken returns the AccessToken field value if set, zero value otherwise.
 func (o *OAuth2TokenExchange) GetAccessToken() string {
-	if o == nil || o.AccessToken == nil {
+	if o == nil || IsNil(o.AccessToken) {
 		var ret string
 		return ret
 	}
@@ -60,7 +63,7 @@ func (o *OAuth2TokenExchange) GetAccessToken() string {
 // GetAccessTokenOk returns a tuple with the AccessToken field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2TokenExchange) GetAccessTokenOk() (*string, bool) {
-	if o == nil || o.AccessToken == nil {
+	if o == nil || IsNil(o.AccessToken) {
 		return nil, false
 	}
 	return o.AccessToken, true
@@ -68,7 +71,7 @@ func (o *OAuth2TokenExchange) GetAccessTokenOk() (*string, bool) {
 
 // HasAccessToken returns a boolean if a field has been set.
 func (o *OAuth2TokenExchange) HasAccessToken() bool {
-	if o != nil && o.AccessToken != nil {
+	if o != nil && !IsNil(o.AccessToken) {
 		return true
 	}
 
@@ -82,7 +85,7 @@ func (o *OAuth2TokenExchange) SetAccessToken(v string) {
 
 // GetExpiresIn returns the ExpiresIn field value if set, zero value otherwise.
 func (o *OAuth2TokenExchange) GetExpiresIn() int64 {
-	if o == nil || o.ExpiresIn == nil {
+	if o == nil || IsNil(o.ExpiresIn) {
 		var ret int64
 		return ret
 	}
@@ -92,7 +95,7 @@ func (o *OAuth2TokenExchange) GetExpiresIn() int64 {
 // GetExpiresInOk returns a tuple with the ExpiresIn field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2TokenExchange) GetExpiresInOk() (*int64, bool) {
-	if o == nil || o.ExpiresIn == nil {
+	if o == nil || IsNil(o.ExpiresIn) {
 		return nil, false
 	}
 	return o.ExpiresIn, true
@@ -100,7 +103,7 @@ func (o *OAuth2TokenExchange) GetExpiresInOk() (*int64, bool) {
 
 // HasExpiresIn returns a boolean if a field has been set.
 func (o *OAuth2TokenExchange) HasExpiresIn() bool {
-	if o != nil && o.ExpiresIn != nil {
+	if o != nil && !IsNil(o.ExpiresIn) {
 		return true
 	}
 
@@ -114,7 +117,7 @@ func (o *OAuth2TokenExchange) SetExpiresIn(v int64) {
 
 // GetIdToken returns the IdToken field value if set, zero value otherwise.
 func (o *OAuth2TokenExchange) GetIdToken() string {
-	if o == nil || o.IdToken == nil {
+	if o == nil || IsNil(o.IdToken) {
 		var ret string
 		return ret
 	}
@@ -124,7 +127,7 @@ func (o *OAuth2TokenExchange) GetIdToken() string {
 // GetIdTokenOk returns a tuple with the IdToken field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2TokenExchange) GetIdTokenOk() (*string, bool) {
-	if o == nil || o.IdToken == nil {
+	if o == nil || IsNil(o.IdToken) {
 		return nil, false
 	}
 	return o.IdToken, true
@@ -132,7 +135,7 @@ func (o *OAuth2TokenExchange) GetIdTokenOk() (*string, bool) {
 
 // HasIdToken returns a boolean if a field has been set.
 func (o *OAuth2TokenExchange) HasIdToken() bool {
-	if o != nil && o.IdToken != nil {
+	if o != nil && !IsNil(o.IdToken) {
 		return true
 	}
 
@@ -146,7 +149,7 @@ func (o *OAuth2TokenExchange) SetIdToken(v string) {
 
 // GetRefreshToken returns the RefreshToken field value if set, zero value otherwise.
 func (o *OAuth2TokenExchange) GetRefreshToken() string {
-	if o == nil || o.RefreshToken == nil {
+	if o == nil || IsNil(o.RefreshToken) {
 		var ret string
 		return ret
 	}
@@ -156,7 +159,7 @@ func (o *OAuth2TokenExchange) GetRefreshToken() string {
 // GetRefreshTokenOk returns a tuple with the RefreshToken field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2TokenExchange) GetRefreshTokenOk() (*string, bool) {
-	if o == nil || o.RefreshToken == nil {
+	if o == nil || IsNil(o.RefreshToken) {
 		return nil, false
 	}
 	return o.RefreshToken, true
@@ -164,7 +167,7 @@ func (o *OAuth2TokenExchange) GetRefreshTokenOk() (*string, bool) {
 
 // HasRefreshToken returns a boolean if a field has been set.
 func (o *OAuth2TokenExchange) HasRefreshToken() bool {
-	if o != nil && o.RefreshToken != nil {
+	if o != nil && !IsNil(o.RefreshToken) {
 		return true
 	}
 
@@ -178,7 +181,7 @@ func (o *OAuth2TokenExchange) SetRefreshToken(v string) {
 
 // GetScope returns the Scope field value if set, zero value otherwise.
 func (o *OAuth2TokenExchange) GetScope() string {
-	if o == nil || o.Scope == nil {
+	if o == nil || IsNil(o.Scope) {
 		var ret string
 		return ret
 	}
@@ -188,7 +191,7 @@ func (o *OAuth2TokenExchange) GetScope() string {
 // GetScopeOk returns a tuple with the Scope field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2TokenExchange) GetScopeOk() (*string, bool) {
-	if o == nil || o.Scope == nil {
+	if o == nil || IsNil(o.Scope) {
 		return nil, false
 	}
 	return o.Scope, true
@@ -196,7 +199,7 @@ func (o *OAuth2TokenExchange) GetScopeOk() (*string, bool) {
 
 // HasScope returns a boolean if a field has been set.
 func (o *OAuth2TokenExchange) HasScope() bool {
-	if o != nil && o.Scope != nil {
+	if o != nil && !IsNil(o.Scope) {
 		return true
 	}
 
@@ -210,7 +213,7 @@ func (o *OAuth2TokenExchange) SetScope(v string) {
 
 // GetTokenType returns the TokenType field value if set, zero value otherwise.
 func (o *OAuth2TokenExchange) GetTokenType() string {
-	if o == nil || o.TokenType == nil {
+	if o == nil || IsNil(o.TokenType) {
 		var ret string
 		return ret
 	}
@@ -220,7 +223,7 @@ func (o *OAuth2TokenExchange) GetTokenType() string {
 // GetTokenTypeOk returns a tuple with the TokenType field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *OAuth2TokenExchange) GetTokenTypeOk() (*string, bool) {
-	if o == nil || o.TokenType == nil {
+	if o == nil || IsNil(o.TokenType) {
 		return nil, false
 	}
 	return o.TokenType, true
@@ -228,7 +231,7 @@ func (o *OAuth2TokenExchange) GetTokenTypeOk() (*string, bool) {
 
 // HasTokenType returns a boolean if a field has been set.
 func (o *OAuth2TokenExchange) HasTokenType() bool {
-	if o != nil && o.TokenType != nil {
+	if o != nil && !IsNil(o.TokenType) {
 		return true
 	}
 
@@ -241,26 +244,34 @@ func (o *OAuth2TokenExchange) SetTokenType(v string) {
 }
 
 func (o OAuth2TokenExchange) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.AccessToken != nil {
-		toSerialize["access_token"] = o.AccessToken
-	}
-	if o.ExpiresIn != nil {
-		toSerialize["expires_in"] = o.ExpiresIn
-	}
-	if o.IdToken != nil {
-		toSerialize["id_token"] = o.IdToken
-	}
-	if o.RefreshToken != nil {
-		toSerialize["refresh_token"] = o.RefreshToken
-	}
-	if o.Scope != nil {
-		toSerialize["scope"] = o.Scope
-	}
-	if o.TokenType != nil {
-		toSerialize["token_type"] = o.TokenType
+	toSerialize, err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o OAuth2TokenExchange) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.AccessToken) {
+		toSerialize["access_token"] = o.AccessToken
+	}
+	if !IsNil(o.ExpiresIn) {
+		toSerialize["expires_in"] = o.ExpiresIn
+	}
+	if !IsNil(o.IdToken) {
+		toSerialize["id_token"] = o.IdToken
+	}
+	if !IsNil(o.RefreshToken) {
+		toSerialize["refresh_token"] = o.RefreshToken
+	}
+	if !IsNil(o.Scope) {
+		toSerialize["scope"] = o.Scope
+	}
+	if !IsNil(o.TokenType) {
+		toSerialize["token_type"] = o.TokenType
+	}
+	return toSerialize, nil
 }
 
 type NullableOAuth2TokenExchange struct {
