@@ -15,6 +15,9 @@ import (
 	"encoding/json"
 )
 
+// checks if the Pagination type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &Pagination{}
+
 // Pagination struct for Pagination
 type Pagination struct {
 	// Items per page  This is the number of items per page to return. For details on pagination please head over to the [pagination documentation](https://www.ory.sh/docs/ecosystem/api-design#pagination).
@@ -50,7 +53,7 @@ func NewPaginationWithDefaults() *Pagination {
 
 // GetPageSize returns the PageSize field value if set, zero value otherwise.
 func (o *Pagination) GetPageSize() int64 {
-	if o == nil || o.PageSize == nil {
+	if o == nil || IsNil(o.PageSize) {
 		var ret int64
 		return ret
 	}
@@ -60,7 +63,7 @@ func (o *Pagination) GetPageSize() int64 {
 // GetPageSizeOk returns a tuple with the PageSize field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Pagination) GetPageSizeOk() (*int64, bool) {
-	if o == nil || o.PageSize == nil {
+	if o == nil || IsNil(o.PageSize) {
 		return nil, false
 	}
 	return o.PageSize, true
@@ -68,7 +71,7 @@ func (o *Pagination) GetPageSizeOk() (*int64, bool) {
 
 // HasPageSize returns a boolean if a field has been set.
 func (o *Pagination) HasPageSize() bool {
-	if o != nil && o.PageSize != nil {
+	if o != nil && !IsNil(o.PageSize) {
 		return true
 	}
 
@@ -82,7 +85,7 @@ func (o *Pagination) SetPageSize(v int64) {
 
 // GetPageToken returns the PageToken field value if set, zero value otherwise.
 func (o *Pagination) GetPageToken() string {
-	if o == nil || o.PageToken == nil {
+	if o == nil || IsNil(o.PageToken) {
 		var ret string
 		return ret
 	}
@@ -92,7 +95,7 @@ func (o *Pagination) GetPageToken() string {
 // GetPageTokenOk returns a tuple with the PageToken field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Pagination) GetPageTokenOk() (*string, bool) {
-	if o == nil || o.PageToken == nil {
+	if o == nil || IsNil(o.PageToken) {
 		return nil, false
 	}
 	return o.PageToken, true
@@ -100,7 +103,7 @@ func (o *Pagination) GetPageTokenOk() (*string, bool) {
 
 // HasPageToken returns a boolean if a field has been set.
 func (o *Pagination) HasPageToken() bool {
-	if o != nil && o.PageToken != nil {
+	if o != nil && !IsNil(o.PageToken) {
 		return true
 	}
 
@@ -113,14 +116,22 @@ func (o *Pagination) SetPageToken(v string) {
 }
 
 func (o Pagination) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.PageSize != nil {
-		toSerialize["page_size"] = o.PageSize
-	}
-	if o.PageToken != nil {
-		toSerialize["page_token"] = o.PageToken
+	toSerialize, err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o Pagination) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.PageSize) {
+		toSerialize["page_size"] = o.PageSize
+	}
+	if !IsNil(o.PageToken) {
+		toSerialize["page_token"] = o.PageToken
+	}
+	return toSerialize, nil
 }
 
 type NullablePagination struct {
