@@ -12,9 +12,14 @@ Contact: hi@ory.sh
 package openapi
 
 import (
+	"bytes"
 	"encoding/json"
+	"fmt"
 	"time"
 )
+
+// checks if the TrustOAuth2JwtGrantIssuer type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &TrustOAuth2JwtGrantIssuer{}
 
 // TrustOAuth2JwtGrantIssuer Trust OAuth2 JWT Bearer Grant Type Issuer Request Body
 type TrustOAuth2JwtGrantIssuer struct {
@@ -30,6 +35,8 @@ type TrustOAuth2JwtGrantIssuer struct {
 	// The \"subject\" identifies the principal that is the subject of the JWT.
 	Subject *string `json:"subject,omitempty"`
 }
+
+type _TrustOAuth2JwtGrantIssuer TrustOAuth2JwtGrantIssuer
 
 // NewTrustOAuth2JwtGrantIssuer instantiates a new TrustOAuth2JwtGrantIssuer object
 // This constructor will assign default values to properties that have it defined,
@@ -54,7 +61,7 @@ func NewTrustOAuth2JwtGrantIssuerWithDefaults() *TrustOAuth2JwtGrantIssuer {
 
 // GetAllowAnySubject returns the AllowAnySubject field value if set, zero value otherwise.
 func (o *TrustOAuth2JwtGrantIssuer) GetAllowAnySubject() bool {
-	if o == nil || o.AllowAnySubject == nil {
+	if o == nil || IsNil(o.AllowAnySubject) {
 		var ret bool
 		return ret
 	}
@@ -64,7 +71,7 @@ func (o *TrustOAuth2JwtGrantIssuer) GetAllowAnySubject() bool {
 // GetAllowAnySubjectOk returns a tuple with the AllowAnySubject field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *TrustOAuth2JwtGrantIssuer) GetAllowAnySubjectOk() (*bool, bool) {
-	if o == nil || o.AllowAnySubject == nil {
+	if o == nil || IsNil(o.AllowAnySubject) {
 		return nil, false
 	}
 	return o.AllowAnySubject, true
@@ -72,7 +79,7 @@ func (o *TrustOAuth2JwtGrantIssuer) GetAllowAnySubjectOk() (*bool, bool) {
 
 // HasAllowAnySubject returns a boolean if a field has been set.
 func (o *TrustOAuth2JwtGrantIssuer) HasAllowAnySubject() bool {
-	if o != nil && o.AllowAnySubject != nil {
+	if o != nil && !IsNil(o.AllowAnySubject) {
 		return true
 	}
 
@@ -182,7 +189,7 @@ func (o *TrustOAuth2JwtGrantIssuer) SetScope(v []string) {
 
 // GetSubject returns the Subject field value if set, zero value otherwise.
 func (o *TrustOAuth2JwtGrantIssuer) GetSubject() string {
-	if o == nil || o.Subject == nil {
+	if o == nil || IsNil(o.Subject) {
 		var ret string
 		return ret
 	}
@@ -192,7 +199,7 @@ func (o *TrustOAuth2JwtGrantIssuer) GetSubject() string {
 // GetSubjectOk returns a tuple with the Subject field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *TrustOAuth2JwtGrantIssuer) GetSubjectOk() (*string, bool) {
-	if o == nil || o.Subject == nil {
+	if o == nil || IsNil(o.Subject) {
 		return nil, false
 	}
 	return o.Subject, true
@@ -200,7 +207,7 @@ func (o *TrustOAuth2JwtGrantIssuer) GetSubjectOk() (*string, bool) {
 
 // HasSubject returns a boolean if a field has been set.
 func (o *TrustOAuth2JwtGrantIssuer) HasSubject() bool {
-	if o != nil && o.Subject != nil {
+	if o != nil && !IsNil(o.Subject) {
 		return true
 	}
 
@@ -213,26 +220,66 @@ func (o *TrustOAuth2JwtGrantIssuer) SetSubject(v string) {
 }
 
 func (o TrustOAuth2JwtGrantIssuer) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.AllowAnySubject != nil {
-		toSerialize["allow_any_subject"] = o.AllowAnySubject
-	}
-	if true {
-		toSerialize["expires_at"] = o.ExpiresAt
-	}
-	if true {
-		toSerialize["issuer"] = o.Issuer
-	}
-	if true {
-		toSerialize["jwk"] = o.Jwk
-	}
-	if true {
-		toSerialize["scope"] = o.Scope
-	}
-	if o.Subject != nil {
-		toSerialize["subject"] = o.Subject
+	toSerialize, err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o TrustOAuth2JwtGrantIssuer) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.AllowAnySubject) {
+		toSerialize["allow_any_subject"] = o.AllowAnySubject
+	}
+	toSerialize["expires_at"] = o.ExpiresAt
+	toSerialize["issuer"] = o.Issuer
+	toSerialize["jwk"] = o.Jwk
+	toSerialize["scope"] = o.Scope
+	if !IsNil(o.Subject) {
+		toSerialize["subject"] = o.Subject
+	}
+	return toSerialize, nil
+}
+
+func (o *TrustOAuth2JwtGrantIssuer) UnmarshalJSON(data []byte) (err error) {
+	// This validates that all required properties are included in the JSON object
+	// by unmarshalling the object into a generic map with string keys and checking
+	// that every required field exists as a key in the generic map.
+	requiredProperties := []string{
+		"expires_at",
+		"issuer",
+		"jwk",
+		"scope",
+	}
+
+	allProperties := make(map[string]interface{})
+
+	err = json.Unmarshal(data, &allProperties)
+
+	if err != nil {
+		return err
+	}
+
+	for _, requiredProperty := range requiredProperties {
+		if _, exists := allProperties[requiredProperty]; !exists {
+			return fmt.Errorf("no value given for required property %v", requiredProperty)
+		}
+	}
+
+	varTrustOAuth2JwtGrantIssuer := _TrustOAuth2JwtGrantIssuer{}
+
+	decoder := json.NewDecoder(bytes.NewReader(data))
+	decoder.DisallowUnknownFields()
+	err = decoder.Decode(&varTrustOAuth2JwtGrantIssuer)
+
+	if err != nil {
+		return err
+	}
+
+	*o = TrustOAuth2JwtGrantIssuer(varTrustOAuth2JwtGrantIssuer)
+
+	return err
 }
 
 type NullableTrustOAuth2JwtGrantIssuer struct {
