@@ -15,6 +15,9 @@ import (
 	"encoding/json"
 )
 
+// checks if the TrustedOAuth2JwtGrantJsonWebKey type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &TrustedOAuth2JwtGrantJsonWebKey{}
+
 // TrustedOAuth2JwtGrantJsonWebKey OAuth2 JWT Bearer Grant Type Issuer Trusted JSON Web Key
 type TrustedOAuth2JwtGrantJsonWebKey struct {
 	// The \"key_id\" is key unique identifier (same as kid header in jws/jwt).
@@ -42,7 +45,7 @@ func NewTrustedOAuth2JwtGrantJsonWebKeyWithDefaults() *TrustedOAuth2JwtGrantJson
 
 // GetKid returns the Kid field value if set, zero value otherwise.
 func (o *TrustedOAuth2JwtGrantJsonWebKey) GetKid() string {
-	if o == nil || o.Kid == nil {
+	if o == nil || IsNil(o.Kid) {
 		var ret string
 		return ret
 	}
@@ -52,7 +55,7 @@ func (o *TrustedOAuth2JwtGrantJsonWebKey) GetKid() string {
 // GetKidOk returns a tuple with the Kid field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *TrustedOAuth2JwtGrantJsonWebKey) GetKidOk() (*string, bool) {
-	if o == nil || o.Kid == nil {
+	if o == nil || IsNil(o.Kid) {
 		return nil, false
 	}
 	return o.Kid, true
@@ -60,7 +63,7 @@ func (o *TrustedOAuth2JwtGrantJsonWebKey) GetKidOk() (*string, bool) {
 
 // HasKid returns a boolean if a field has been set.
 func (o *TrustedOAuth2JwtGrantJsonWebKey) HasKid() bool {
-	if o != nil && o.Kid != nil {
+	if o != nil && !IsNil(o.Kid) {
 		return true
 	}
 
@@ -74,7 +77,7 @@ func (o *TrustedOAuth2JwtGrantJsonWebKey) SetKid(v string) {
 
 // GetSet returns the Set field value if set, zero value otherwise.
 func (o *TrustedOAuth2JwtGrantJsonWebKey) GetSet() string {
-	if o == nil || o.Set == nil {
+	if o == nil || IsNil(o.Set) {
 		var ret string
 		return ret
 	}
@@ -84,7 +87,7 @@ func (o *TrustedOAuth2JwtGrantJsonWebKey) GetSet() string {
 // GetSetOk returns a tuple with the Set field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *TrustedOAuth2JwtGrantJsonWebKey) GetSetOk() (*string, bool) {
-	if o == nil || o.Set == nil {
+	if o == nil || IsNil(o.Set) {
 		return nil, false
 	}
 	return o.Set, true
@@ -92,7 +95,7 @@ func (o *TrustedOAuth2JwtGrantJsonWebKey) GetSetOk() (*string, bool) {
 
 // HasSet returns a boolean if a field has been set.
 func (o *TrustedOAuth2JwtGrantJsonWebKey) HasSet() bool {
-	if o != nil && o.Set != nil {
+	if o != nil && !IsNil(o.Set) {
 		return true
 	}
 
@@ -105,14 +108,22 @@ func (o *TrustedOAuth2JwtGrantJsonWebKey) SetSet(v string) {
 }
 
 func (o TrustedOAuth2JwtGrantJsonWebKey) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.Kid != nil {
-		toSerialize["kid"] = o.Kid
-	}
-	if o.Set != nil {
-		toSerialize["set"] = o.Set
+	toSerialize, err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o TrustedOAuth2JwtGrantJsonWebKey) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.Kid) {
+		toSerialize["kid"] = o.Kid
+	}
+	if !IsNil(o.Set) {
+		toSerialize["set"] = o.Set
+	}
+	return toSerialize, nil
 }
 
 type NullableTrustedOAuth2JwtGrantJsonWebKey struct {
